refactor: split converter.py into focused internal modules
Break the single 2,980-line converter.py into ten cohesive internal
modules with an acyclic dependency graph:

  _constants     shared constants, regexes, measured Slack limits
  _code_regions  fence + inline-code-span tracking (single source of truth)
  _sanitize      entities, control chars, angle tokens, bare-URL autolinks
  _emphasis      underscore normalization + ZWSP/visible-space engine
  _lines         line classifiers, list context, blank-line placeholders
  _rich_text     rich_text inline elements and plain-text downgrades
  _tables        table normalization, cell splitting, table block output
  _splitting     size splitting, expansion estimation, message packing
  _rich_blocks   promotion of standalone constructs to native blocks
  _fallback      preview / fallback text generation

Code was moved verbatim: every definition is AST-identical to the
previous converter.py, so behavior is unchanged (207 tests pass
unmodified). converter.py keeps the pipeline orchestration and
re-exports every previously importable name — public and private —
so slack_markdown_parser.converter imports keep working.

Add ruff isort combine-as-imports so the facade's re-export block
stays grouped one statement per module.

Motivation: duplicated-logic drift across passes was the root cause
of the 2.5.0 sanitize corruption; module boundaries now work against
that class of bug.
