+++ pyproject.toml
@@ -58,5 +58,8 @@ target-version = "py310"
 select = ["E", "F", "I", "UP", "B"]
 ignore = ["E501"]  # line length is enforced by black
 
+[tool.ruff.lint.isort]
+combine-as-imports = true  # keep the converter facade's re-exports grouped per module
+
 [tool.pytest.ini_options]
 testpaths = ["tests"]
