+++ slack_markdown_parser/converter.py
@@ -1,2590 +1,180 @@
-"""Core conversion utilities for Slack Block Kit output.
+"""Core conversion pipeline for Slack Block Kit output.
 
-This module converts LLM-generated Markdown text into Slack Block Kit blocks,
-with support for Slack table blocks and robust fallback text generation.
-"""
+This module orchestrates the conversion of LLM-generated Markdown text into
+Slack Block Kit blocks. The implementation lives in focused internal modules
+(``_sanitize``, ``_emphasis``, ``_tables``, ``_rich_blocks``, ``_splitting``,
+...); every name that was historically importable from
+``slack_markdown_parser.converter`` is re-exported here so existing imports
+keep working."""
 
 from __future__ import annotations
 
-import html
-import re
-from collections.abc import Callable, Iterable, Iterator
 from typing import Any
-from urllib.parse import urlparse
 
-ZWSP = "\u200b"
-NBSP = "\u00a0"
-VISIBLE_BOUNDARY_CHARS = {" ", "\t", "\n", "\r"}
-SYNTH_SPACE_MARKER = "\u2063"
-STRIP_LEFT_ZWSP_MARKER = "\ufff2"
-STRIP_RIGHT_ZWSP_MARKER = "\ufff3"
-
-ANSI_ESCAPE_PATTERN = re.compile(
-    r"\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~]|\][^\x1B\x07]*(?:\x07|\x1B\\))"
+# Backward-compatible re-exports: everything that used to be defined in this
+# module remains importable from ``slack_markdown_parser.converter``. The
+# orchestration functions below use a subset of these names directly.
+from ._code_regions import (
+    _find_inline_code_span_end as _find_inline_code_span_end,
+    _is_fence_close as _is_fence_close,
+    _iter_fence_states as _iter_fence_states,
+    _iter_inline_code_spans as _iter_inline_code_spans,
+    _match_fence_open as _match_fence_open,
+    _multiline_code_span_line_map as _multiline_code_span_line_map,
+    _split_fenced_code_chunks as _split_fenced_code_chunks,
+    _transform_outside_code_regions as _transform_outside_code_regions,
+    _transform_outside_inline_code as _transform_outside_inline_code,
 )
-CONTROL_CHAR_PATTERN = re.compile(r"[\x00-\x08\x0B\x0C\x0E-\x1F\x7F-\x9F]")
-# In-band marker code points reserved by this module's placeholder/spacing
-# machinery: SYNTH_SPACE_MARKER (U+2063), the inline-code placeholder
-# delimiters (U+FFF0/U+FFF1), and the ZWSP-strip markers (U+FFF2/U+FFF3).
-# They have no legitimate use in chat text, and input that carries them would
-# collide with internal placeholders (a stray ``\ufff0code0\ufff1`` either
-# crashes restoration with KeyError or gets substituted with another span's
-# content), so they are removed up front together with control characters.
-INTERNAL_MARKER_CHAR_PATTERN = re.compile("[\u2063\ufff0-\ufff3]")
-SLACK_ANGLE_TOKEN_PATTERN = re.compile(r"<[^>\n]+>")
-BARE_URL_PATTERN = re.compile(r"https?://[^\s<]+", re.IGNORECASE)
-FENCE_OPEN_PATTERN = re.compile(r"^[ \t]{0,3}(`{3,}|~{3,})([^\n]*)$")
-STANDALONE_IMAGE_PATTERN = re.compile(
-    r"^[ \t]*!\[(?P<alt>[^\]\n]*)\]\((?P<url>https?://[^\s)]+)"
-    r"(?:[ \t]+(?P<title>\"[^\"\n]*\"|'[^'\n]*'))?[ \t]*\)[ \t]*$",
-    re.IGNORECASE,
+from ._constants import (
+    _CODE_SPAN_BLANK_LINE_PATTERN as _CODE_SPAN_BLANK_LINE_PATTERN,
+    _MARKDOWN_EXPANSION_ITEMS_TARGET as _MARKDOWN_EXPANSION_ITEMS_TARGET,
+    _MARKDOWN_SPLIT_TARGET_LENGTH as _MARKDOWN_SPLIT_TARGET_LENGTH,
+    _MESSAGE_BLOCKS_TEXT_TARGET as _MESSAGE_BLOCKS_TEXT_TARGET,
+    _QUOTE_MARKER_PREFIX_PATTERN as _QUOTE_MARKER_PREFIX_PATTERN,
+    _TEXT_SIZE_KEYS as _TEXT_SIZE_KEYS,
+    _URL_CJK_BOUNDARY_CHARS as _URL_CJK_BOUNDARY_CHARS,
+    _URL_STOP_CHARS as _URL_STOP_CHARS,
+    _URL_TRAILING_PUNCTUATION as _URL_TRAILING_PUNCTUATION,
+    ALLOWED_SLACK_ANGLE_TOKEN_PATTERNS as ALLOWED_SLACK_ANGLE_TOKEN_PATTERNS,
+    ANSI_ESCAPE_PATTERN as ANSI_ESCAPE_PATTERN,
+    ATX_HEADING_PATTERN as ATX_HEADING_PATTERN,
+    BARE_URL_PATTERN as BARE_URL_PATTERN,
+    CONTROL_CHAR_PATTERN as CONTROL_CHAR_PATTERN,
+    DOUBLE_UNDERSCORE_EMPHASIS_PATTERN as DOUBLE_UNDERSCORE_EMPHASIS_PATTERN,
+    EMPHASIS_PATTERNS as EMPHASIS_PATTERNS,
+    FENCE_OPEN_PATTERN as FENCE_OPEN_PATTERN,
+    INLINE_CODE_PLACEHOLDER_PATTERN as INLINE_CODE_PLACEHOLDER_PATTERN,
+    INTERNAL_MARKER_CHAR_PATTERN as INTERNAL_MARKER_CHAR_PATTERN,
+    LIST_ITEM_PATTERN as LIST_ITEM_PATTERN,
+    LOOSE_TABLE_SEPARATOR_PATTERN as LOOSE_TABLE_SEPARATOR_PATTERN,
+    MARKDOWN_BACKSLASH_ESCAPE_PATTERN as MARKDOWN_BACKSLASH_ESCAPE_PATTERN,
+    MARKDOWN_LINK_PATTERN as MARKDOWN_LINK_PATTERN,
+    NBSP as NBSP,
+    PROTECTED_UNDERSCORE_SPAN_PATTERN as PROTECTED_UNDERSCORE_SPAN_PATTERN,
+    REFERENCE_DEFINITION_PATTERN as REFERENCE_DEFINITION_PATTERN,
+    SETEXT_HEADING_UNDERLINE_PATTERN as SETEXT_HEADING_UNDERLINE_PATTERN,
+    SINGLE_UNDERSCORE_EMPHASIS_PATTERN as SINGLE_UNDERSCORE_EMPHASIS_PATTERN,
+    SLACK_ANGLE_TOKEN_PATTERN as SLACK_ANGLE_TOKEN_PATTERN,
+    SLACK_MAX_BLOCKS_PER_MESSAGE as SLACK_MAX_BLOCKS_PER_MESSAGE,
+    SLACK_MAX_EXPANSION_ITEMS_PER_MESSAGE as SLACK_MAX_EXPANSION_ITEMS_PER_MESSAGE,
+    SLACK_MAX_MARKDOWN_TEXT_LENGTH as SLACK_MAX_MARKDOWN_TEXT_LENGTH,
+    SLACK_MAX_MESSAGE_BLOCKS_TEXT_LENGTH as SLACK_MAX_MESSAGE_BLOCKS_TEXT_LENGTH,
+    STANDALONE_IMAGE_PATTERN as STANDALONE_IMAGE_PATTERN,
+    STRIP_LEFT_ZWSP_MARKER as STRIP_LEFT_ZWSP_MARKER,
+    STRIP_RIGHT_ZWSP_MARKER as STRIP_RIGHT_ZWSP_MARKER,
+    SYNTH_SPACE_MARKER as SYNTH_SPACE_MARKER,
+    TABLE_SEPARATOR_PATTERN as TABLE_SEPARATOR_PATTERN,
+    TABLE_TOKEN_PATTERN as TABLE_TOKEN_PATTERN,
+    TASK_LIST_MARKER_PATTERN as TASK_LIST_MARKER_PATTERN,
+    THEMATIC_BREAK_PATTERN as THEMATIC_BREAK_PATTERN,
+    VISIBLE_BOUNDARY_CHARS as VISIBLE_BOUNDARY_CHARS,
+    ZWSP as ZWSP,
 )
-MARKDOWN_LINK_PATTERN = re.compile(r"\[[^\]\n]+\]\([^\)\n]+\)")
-# Emphasis delimiters must satisfy CommonMark's minimal flanking requirement:
-# an opening run is not followed by whitespace and a closing run is not preceded
-# by whitespace. Enforcing this keeps a stray, whitespace-flanked delimiter
-# (e.g. the literal ``**`` in ``閉じ ** が``) from being paired at all.
-#
-# For ``**`` and ``~~`` the body additionally may not contain the same delimiter
-# run (``(?:(?!\*\*).)+?`` / ``(?:(?!~~).)+?``). Without this, a dangling opener
-# with no valid closer of its own (``**oops ** and **70.9%→83.0%**``) would scan
-# past the literal stray and steal a *later* well-formed span's closing marker,
-# shifting the pairing and corrupting that span's ZWSP placement. Bounding the
-# body to a single run makes the regex pair the same markers CommonMark does.
-# (The single-``*`` italic body is intentionally not bounded this way: italics
-# legitimately wrap ``**bold**`` and ``*`` is heavily overloaded, so it keeps the
-# whitespace guard only.)
-#
-# Every body additionally may not cross a blank line (``(?!\n[ \t\r]*\n)``,
-# CRLF included): CommonMark emphasis never spans paragraphs, so a stray
-# ``*``/``**``/``~~`` in one paragraph must not pair with a stray marker in a
-# later paragraph and get ZWSP-padded as though it were one span.
-EMPHASIS_PATTERNS = (
-    re.compile(
-        r"(?<!\*)\*\*(?!\s)((?:(?!\*\*|\n[ \t\r]*\n).)+?)(?<!\s)\*\*(?!\*)",
-        flags=re.DOTALL,
-    ),
-    re.compile(
-        r"(?<!\*)\*(?!\*)(?!\s)((?:(?!\n[ \t\r]*\n).)+?)(?<!\s)(?<!\*)\*(?!\*)",
-        flags=re.DOTALL,
-    ),
-    re.compile(r"~~(?!\s)((?:(?!~~|\n[ \t\r]*\n).)+?)(?<!\s)~~", flags=re.DOTALL),
+from ._emphasis import (
+    _format_markdown_with_spacing_metadata as _format_markdown_with_spacing_metadata,
+    _is_han_or_kana_char as _is_han_or_kana_char,
+    _is_hangul_char as _is_hangul_char,
+    _is_punctuation_like as _is_punctuation_like,
+    _needs_inner_code_spacing as _needs_inner_code_spacing,
+    _nested_code_space_strategy as _nested_code_space_strategy,
+    _normalize_underscore_emphasis_chunk as _normalize_underscore_emphasis_chunk,
+    _normalize_underscore_emphasis_prose as _normalize_underscore_emphasis_prose,
+    _remove_synthetic_space_markers as _remove_synthetic_space_markers,
+    _should_preserve_raw_punctuation_emphasis as _should_preserve_raw_punctuation_emphasis,
+    add_zero_width_spaces as add_zero_width_spaces,
+    add_zero_width_spaces_to_markdown as add_zero_width_spaces_to_markdown,
+    normalize_underscore_emphasis as normalize_underscore_emphasis,
 )
-INLINE_CODE_PLACEHOLDER_PATTERN = re.compile(r"\ufff0code\d+\ufff1")
-# A blank line ends the paragraph, and with it any possible code span: Slack
-# pairs backticks across soft line breaks but never across paragraphs. The
-# ``\r`` in the class keeps CRLF blank lines (``\r\n\r\n``) recognized too.
-_CODE_SPAN_BLANK_LINE_PATTERN = re.compile(r"\n[ \t\r]*\n")
-PROTECTED_UNDERSCORE_SPAN_PATTERN = re.compile(
-    r"`[^`\n]+`"
-    r"|\[[^\]\n]+\]\([^\)\n]+\)"
-    r"|<[^>\n]+>"
-    r"|https?://[^\s<]+"
-    r"|mailto:[^\s<]+",
-    re.IGNORECASE,
+from ._fallback import (
+    _blocks_to_downgrade_parts as _blocks_to_downgrade_parts,
+    _build_markdown_block_plain_text as _build_markdown_block_plain_text,
+    _markdown_block_to_plain_text as _markdown_block_to_plain_text,
+    _normalize_markdown_block_plain_text as _normalize_markdown_block_plain_text,
+    _strip_synthetic_spaces_from_plain_text as _strip_synthetic_spaces_from_plain_text,
+    blocks_to_plain_text as blocks_to_plain_text,
+    build_fallback_text_from_blocks as build_fallback_text_from_blocks,
 )
-REFERENCE_DEFINITION_PATTERN = re.compile(r"^[ \t]{0,3}\[[^\]\n]+\]:")
-SETEXT_HEADING_UNDERLINE_PATTERN = re.compile(r"^[ \t]{0,3}(?:=+|-+)\s*$")
-THEMATIC_BREAK_PATTERN = re.compile(
-    r"^[ \t]{0,3}(?P<char>[-_*])(?:[ \t]*\1){2,}[ \t]*$"
+from ._lines import (
+    _blank_run_follows_list_context as _blank_run_follows_list_context,
+    _has_markdown_backslash_escape as _has_markdown_backslash_escape,
+    _indent_width as _indent_width,
+    _inject_visual_blank_line_placeholders as _inject_visual_blank_line_placeholders,
+    _inject_visual_blank_line_placeholders_in_chunk as _inject_visual_blank_line_placeholders_in_chunk,
+    _is_ambiguous_rich_list_indent as _is_ambiguous_rich_list_indent,
+    _is_ordered_list_marker as _is_ordered_list_marker,
+    _is_thematic_break_line as _is_thematic_break_line,
+    _line_belongs_to_list_context as _line_belongs_to_list_context,
+    _list_indent_level as _list_indent_level,
+    _list_item_content_indent as _list_item_content_indent,
+    _ordered_list_marker_number as _ordered_list_marker_number,
+    _ordered_list_marker_starts_at_one as _ordered_list_marker_starts_at_one,
+    _starts_root_list_item as _starts_root_list_item,
+    _strip_synthetic_blank_line_placeholders as _strip_synthetic_blank_line_placeholders,
 )
-LIST_ITEM_PATTERN = re.compile(
-    r"^(?P<indent>[ \t]*)(?P<marker>\d+[.)]|[-+*])(?P<spacing>[ \t]+|$)"
+from ._rich_blocks import (
+    _consume_fenced_code_block as _consume_fenced_code_block,
+    _consume_list_block as _consume_list_block,
+    _consume_quote_block as _consume_quote_block,
+    _consume_rich_markdown_block as _consume_rich_markdown_block,
+    _create_divider_block_from_line as _create_divider_block_from_line,
+    _create_image_block_from_line as _create_image_block_from_line,
+    _create_rich_text_block as _create_rich_text_block,
+    _find_fence_close_index as _find_fence_close_index,
+    _is_http_url as _is_http_url,
+    _is_setext_heading_underline as _is_setext_heading_underline,
+    _parse_simple_list_item as _parse_simple_list_item,
+    _quote_lines_are_simple as _quote_lines_are_simple,
+    _strip_quote_marker as _strip_quote_marker,
+    _truncate_plain_text as _truncate_plain_text,
 )
-TASK_LIST_MARKER_PATTERN = re.compile(r"^\[[ xX]\](?:[ \t]+|$)")
-MARKDOWN_BACKSLASH_ESCAPE_PATTERN = re.compile(r"\\[\\`*_{}\[\]()#+\-.!>|]")
-DOUBLE_UNDERSCORE_EMPHASIS_PATTERN = re.compile(
-    r"(?<![\\0-9A-Za-z_])__(?=\S)(.+?\S)__(?![0-9A-Za-z_])"
+from ._rich_text import (
+    _create_rich_text_inline_elements as _create_rich_text_inline_elements,
+    _create_rich_text_section as _create_rich_text_section,
+    _plain_text_from_markdown_fragment as _plain_text_from_markdown_fragment,
+    _rich_text_block_to_plain_text as _rich_text_block_to_plain_text,
+    _rich_text_inline_elements_to_plain_text as _rich_text_inline_elements_to_plain_text,
+    _rich_text_object_to_plain_text as _rich_text_object_to_plain_text,
+    _slack_mention_element as _slack_mention_element,
+    _slack_mention_element_to_token as _slack_mention_element_to_token,
 )
-SINGLE_UNDERSCORE_EMPHASIS_PATTERN = re.compile(
-    r"(?<![\\0-9A-Za-z_])_(?!_)(?=\S)(.+?\S)_(?![0-9A-Za-z_])"
+from ._sanitize import (
+    _is_allowed_slack_angle_token as _is_allowed_slack_angle_token,
+    _is_url_boundary_char as _is_url_boundary_char,
+    _match_slack_angle_token_end as _match_slack_angle_token_end,
+    _trim_bare_url as _trim_bare_url,
+    decode_html_entities as decode_html_entities,
+    normalize_bare_urls_for_slack_markdown as normalize_bare_urls_for_slack_markdown,
+    sanitize_slack_text as sanitize_slack_text,
+    strip_zero_width_spaces as strip_zero_width_spaces,
 )
-
-TABLE_SEPARATOR_PATTERN = re.compile(r"^\s*\|[\s\-:|]+\|\s*$")
-LOOSE_TABLE_SEPARATOR_PATTERN = re.compile(
-    r"^\s*\|?\s*:?-{3,}\s*(\|\s*:?-{3,}\s*)+\|?\s*$"
+from ._splitting import (
+    _AnnotatedSlackBlock as _AnnotatedSlackBlock,
+    _block_expansion_weight as _block_expansion_weight,
+    _block_text_size as _block_text_size,
+    _create_markdown_block as _create_markdown_block,
+    _create_markdown_blocks as _create_markdown_blocks,
+    _estimate_markdown_expansion_items as _estimate_markdown_expansion_items,
+    _is_markdown_expansion_breaker_line as _is_markdown_expansion_breaker_line,
+    _markdown_block_fits_slack_limits as _markdown_block_fits_slack_limits,
+    _split_line_verbatim_to_length as _split_line_verbatim_to_length,
+    _split_lines_to_length as _split_lines_to_length,
+    _split_single_line_to_length as _split_single_line_to_length,
+    _split_text_at_blank_lines as _split_text_at_blank_lines,
+    split_blocks_by_table as split_blocks_by_table,
 )
-TABLE_TOKEN_PATTERN = re.compile(
-    r"\[(?P<markdown_label>[^\]\n]+)\]\((?P<markdown_url>https?://[^\s)]+)\)"
-    r"|<(?P<angle_url>https?://[^>\s|]+)(?:\|(?P<angle_label>[^>\n]+))?>"
-    # Slack mention tokens: user (<@U…>/<@W…>), channel (<#C…>/<#G…>), user
-    # group (<!subteam^S…>), and broadcast (<!here>/<!channel>/<!everyone>).
-    # An optional ``|label`` is the human-readable display the author saw; the
-    # rich_text element is rendered by Slack from the id, so the label is dropped.
-    r"|<(?P<mention>@[UW][A-Z0-9]+|#[CG][A-Z0-9]+|!subteam\^[A-Z0-9]+"
-    r"|!(?:here|channel|everyone))(?:\|[^>\n]+)?>"
-    r"|(?P<token>"
-    r"(?P<code>(?P<code_delimiter>`+)(?P<code_text>[^\n]+?)(?P=code_delimiter))"
-    r"|~~[^~]+~~"
-    r"|\*\*[^*]+\*\*"
-    r"|(?<!\*)\*[^*]+\*(?!\*)"
-    r")"
+from ._tables import (
+    _count_cell_words as _count_cell_words,
+    _create_table_cell as _create_table_cell,
+    _looks_like_markdown_table as _looks_like_markdown_table,
+    _split_heading_and_table_row as _split_heading_and_table_row,
+    _split_heading_prefix_and_first_cell as _split_heading_prefix_and_first_cell,
+    _split_markdown_table_cells as _split_markdown_table_cells,
+    extract_plain_text_from_table_cell as extract_plain_text_from_table_cell,
+    markdown_table_to_slack_table as markdown_table_to_slack_table,
+    markdown_table_to_table_block as markdown_table_to_table_block,
+    normalize_markdown_tables as normalize_markdown_tables,
+    parse_markdown_table as parse_markdown_table,
 )
-ALLOWED_SLACK_ANGLE_TOKEN_PATTERNS = (
-    re.compile(r"^<https?://[^>\s|]+(?:\|[^>\n]+)?>$"),
-    re.compile(r"^<mailto:[^>\s|]+(?:\|[^>\n]+)?>$"),
-    re.compile(r"^<@[UW][A-Z0-9]+(?:\|[^>\n]+)?>$"),
-    re.compile(r"^<#[CG][A-Z0-9]+(?:\|[^>\n]+)?>$"),
-    re.compile(r"^<!(?:here|channel|everyone)>$"),
-    re.compile(r"^<!subteam\^[A-Z0-9]+(?:\|[^>\n]+)?>$"),
-    re.compile(r"^<!date\^[^>\n]+>$"),
-)
-SLACK_MAX_BLOCKS_PER_MESSAGE = 50
-# Verified against a real Slack workspace (2026-06-11): a ``markdown`` block's
-# ``text`` accepts exactly 12,000 characters, while 12,001 fails the whole
-# chat.postMessage call with ``msg_too_long``. The top-level fallback ``text``
-# field is not subject to this limit (40,001 characters was accepted).
-SLACK_MAX_MARKDOWN_TEXT_LENGTH = 12000
-# Raw-content packing target used when an oversized markdown segment is split.
-# Formatting inflates text (ZWSP padding, NBSP blank-line placeholders), so
-# pieces are packed below the hard limit; the block builder re-splits any
-# piece whose *formatted* text still exceeds the hard limit.
-_MARKDOWN_SPLIT_TARGET_LENGTH = 11500
-# Slack expands a ``markdown`` block server-side into native blocks and then
-# enforces "no more than 50 items" on the expanded result *per message*.
-# Measured against a real workspace (2026-06-11): each heading and each
-# thematic break becomes its own item (50 headings accepted, 51 rejected;
-# 30 headings in each of two blocks rejected), while paragraphs, lists,
-# quotes, and fenced code merge into one item per run between those breakers
-# (60 blank-separated paragraphs and 52 fences were accepted).
-SLACK_MAX_EXPANSION_ITEMS_PER_MESSAGE = 50
-# Per-block packing target, leaving headroom for estimation error.
-_MARKDOWN_EXPANSION_ITEMS_TARGET = 45
-# Third measured hard limit (2026-06-11): the text carried by one message's
-# blocks *in total* — across block types; rich_text content counts too — may
-# not exceed 13,200 characters (= 1.1 × the single-block limit). 13,201
-# fails the whole call with ``msg_blocks_too_long``.
-SLACK_MAX_MESSAGE_BLOCKS_TEXT_LENGTH = 13200
-# Packing target, leaving headroom for structural fields the size proxy may
-# not count exactly.
-_MESSAGE_BLOCKS_TEXT_TARGET = 12800
-ATX_HEADING_PATTERN = re.compile(r"^[ \t]{0,3}#{1,6}(?:[ \t]|$)")
-
-
-def decode_html_entities(text: str) -> str:
-    """Decode HTML entities in prose while leaving code regions verbatim.
-
-    Prose entities such as ``&gt;`` are decoded for natural reading, but a
-    fenced code block or inline code span showing ``&amp;`` keeps the literal
-    entity: code samples are content, not markup to repair.
-    """
-    if not text or "&" not in text:
-        return text
-    return _transform_outside_code_regions(text, html.unescape)
-
-
-def strip_zero_width_spaces(text: str) -> str:
-    """Strip zero-width spaces from text."""
-    return re.sub(r"[\u200B\uFEFF]", "", text or "")
-
-
-class _AnnotatedSlackBlock(dict):
-    """Dict-like block carrying non-serialized metadata for local helpers."""
-
-
-def _is_hangul_char(char: str) -> bool:
-    if not char:
-        return False
-    codepoint = ord(char)
-    return (
-        0x1100 <= codepoint <= 0x11FF
-        or 0x3130 <= codepoint <= 0x318F
-        or 0xAC00 <= codepoint <= 0xD7AF
-    )
-
-
-def _is_han_or_kana_char(char: str) -> bool:
-    if not char:
-        return False
-    codepoint = ord(char)
-    return (
-        0x3040 <= codepoint <= 0x309F
-        or 0x30A0 <= codepoint <= 0x30FF
-        or 0x3400 <= codepoint <= 0x4DBF
-        or 0x4E00 <= codepoint <= 0x9FFF
-        or 0xF900 <= codepoint <= 0xFAFF
-    )
-
-
-# Code/angle/pipe markers that never appear inside a bare URL in this library's
-# prose context. (A single ``*`` and CJK letters are intentionally NOT here: a
-# URL may legally contain a wildcard/query ``*`` and an IRI/IDN may contain CJK
-# letters, so those must be preserved.)
-_URL_STOP_CHARS = frozenset("`<>|")
-# Trailing punctuation stripped from the end of a bare URL. This is exactly
-# GFM's autolink-extension set (``! ? . , : * _ ~``); a closing paren is handled
-# separately, with balancing. ``;`` and quotes are intentionally NOT included —
-# ``;`` is URL-legal in matrix/path parameters and quotes are sub-delimiters, so
-# trimming them could change the link target rather than just shedding prose.
-_URL_TRAILING_PUNCTUATION = frozenset("!?.,:*_~")
-# CJK and full/half-width punctuation/brackets that terminate prose, so a bare
-# URL is cut here. This is an explicit set rather than the whole U+3000–U+303F
-# block on purpose: letter-like CJK iteration marks (々 U+3005, 〻 U+303B),
-# ditto/closure marks (〆 U+3006) and the ideographic number zero (〇 U+3007)
-# are *excluded* so IRIs such as ``https://ja.wikipedia.org/wiki/人々`` survive.
-_URL_CJK_BOUNDARY_CHARS = frozenset(
-    "、。〃〈〉《》「」『』【】〔〕〖〗〘〙〚〛〜〝〞・…"  # CJK punctuation & brackets
-    "！？，．：；（）［］｛｝＜＞｜"  # full-width punctuation & brackets
-    "｡｢｣､"  # half-width CJK punctuation & brackets
-)
-
-
-def _is_url_boundary_char(char: str) -> bool:
-    """Return True when ``char`` is a hard boundary where a bare URL must stop.
-
-    Only unambiguous prose/markup boundaries qualify: code/angle/pipe markers
-    and CJK/full-width *punctuation* (``、``/``。``/``」``/``）`` …). CJK
-    *letters* (including iteration marks like ``々``) are not a boundary, so
-    IRIs such as ``https://ja.wikipedia.org/wiki/人々`` survive.
-    """
-    return char in _URL_STOP_CHARS or char in _URL_CJK_BOUNDARY_CHARS
-
-
-def _trim_bare_url(url: str) -> str:
-    """Trim a greedily matched bare URL down to its real extent.
-
-    In CJK writing a URL is usually glued directly to the following text with no
-    whitespace, so the greedy ``[^\\s<]+`` match would otherwise swallow the
-    trailing ``)``/``**``/``。`` and the rest of the sentence. This stops the URL
-    at the first hard boundary or doubled emphasis run (``**``/``~~``) — single
-    ``*`` and CJK letters are preserved — then drops GFM-style trailing
-    punctuation and unbalanced closing parens, so ``https://example.com)**。``
-    becomes ``https://example.com``.
-    """
-    for index, char in enumerate(url):
-        nxt = url[index + 1] if index + 1 < len(url) else ""
-        if _is_url_boundary_char(char) or (char in "*~" and nxt == char):
-            url = url[:index]
-            break
-
-    while url:
-        last = url[-1]
-        if last == ")":
-            if url.count(")") <= url.count("("):
-                break
-            url = url[:-1]
-            continue
-        if last in _URL_TRAILING_PUNCTUATION:
-            url = url[:-1]
-            continue
-        break
-
-    return url
-
-
-def _nested_code_space_strategy(
-    source: str,
-    start: int,
-    end: int,
-    boundary_chars: set[str] | None = None,
-) -> str | None:
-    boundary_chars = boundary_chars or {*VISIBLE_BOUNDARY_CHARS, ZWSP}
-    neighbors = []
-
-    left_idx = start - 1
-    while left_idx >= 0 and source[left_idx] in boundary_chars:
-        left_idx -= 1
-    if left_idx >= 0:
-        neighbors.append(source[left_idx])
-
-    right_idx = end
-    while right_idx < len(source) and source[right_idx] in boundary_chars:
-        right_idx += 1
-    if right_idx < len(source):
-        neighbors.append(source[right_idx])
-
-    if any(_is_han_or_kana_char(char) for char in neighbors):
-        return "ja_zh"
-    if any(_is_hangul_char(char) for char in neighbors):
-        return "ko"
-    return None
-
-
-def _needs_inner_code_spacing(char: str, boundary_chars: set[str]) -> bool:
-    return bool(char) and char not in boundary_chars and char.isalnum()
-
-
-def _normalize_markdown_block_plain_text(text: str) -> str:
-    if not text:
-        return text
-
-    return re.sub(r"<(https?://[^>\s|]+)>", r"\1", text)
-
-
-def _build_markdown_block_plain_text(
-    text: str, synthetic_space_indices: list[int] | None = None
-) -> str:
-    """Build fallback/plain text for a markdown block before visual-only rewrites."""
-    return _normalize_markdown_block_plain_text(
-        _strip_synthetic_spaces_from_plain_text(
-            strip_zero_width_spaces(text),
-            synthetic_space_indices,
-        )
-    )
-
-
-def _strip_synthetic_spaces_from_plain_text(
-    text: str, synthetic_space_indices: list[int] | None = None
-) -> str:
-    if not text or not synthetic_space_indices:
-        return text
-
-    indices = set(synthetic_space_indices)
-    return "".join(
-        char for idx, char in enumerate(text) if not (idx in indices and char == " ")
-    )
-
-
-def _indent_width(text: str) -> int:
-    width = 0
-    for char in text:
-        if char == " ":
-            width += 1
-        elif char == "\t":
-            # Treat tabs conservatively for list-continuation heuristics.
-            width += 4
-        else:
-            break
-    return width
-
-
-def _list_item_content_indent(match: re.Match[str]) -> int:
-    spacing = match.group("spacing") or ""
-    spacing_width = _indent_width(spacing) if spacing else 1
-    return (
-        _indent_width(match.group("indent") or "")
-        + len(match.group("marker") or "")
-        + spacing_width
-    )
-
-
-def _is_ordered_list_marker(marker: str) -> bool:
-    return bool(marker) and marker[0].isdigit()
-
-
-def _ordered_list_marker_number(marker: str) -> int:
-    if not _is_ordered_list_marker(marker):
-        return 1
-    try:
-        return int(marker[:-1])
-    except ValueError:
-        return 1
-
-
-def _list_indent_level(indent: str) -> int:
-    width = _indent_width(indent or "")
-    if width <= 3:
-        return 0
-    return min(8, ((width - 4) // 4) + 1)
-
-
-def _is_ambiguous_rich_list_indent(indent: str) -> bool:
-    width = _indent_width(indent or "")
-    return 0 < width <= 3
-
-
-def _has_markdown_backslash_escape(text: str) -> bool:
-    return bool(MARKDOWN_BACKSLASH_ESCAPE_PATTERN.search(text or ""))
-
-
-def _is_thematic_break_line(line: str) -> bool:
-    return bool(THEMATIC_BREAK_PATTERN.match(line))
-
-
-def _ordered_list_marker_starts_at_one(marker: str) -> bool:
-    if not _is_ordered_list_marker(marker):
-        return False
-    return marker[:-1] == "1"
-
-
-def _starts_root_list_item(lines: list[str], marker_index: int) -> bool:
-    line = lines[marker_index]
-    if _is_thematic_break_line(line):
-        return False
-    match = LIST_ITEM_PATTERN.match(line)
-    if not match or _indent_width(match.group("indent") or "") > 3:
-        return False
-
-    marker = match.group("marker") or ""
-    if marker_index == 0 or not lines[marker_index - 1].strip(" \t\r"):
-        return True
-
-    if not _is_ordered_list_marker(marker):
-        return True
-
-    return _ordered_list_marker_starts_at_one(marker)
-
-
-def _line_belongs_to_list_context(
-    lines: list[str], *, marker_index: int, target_index: int
-) -> bool:
-    marker_match = LIST_ITEM_PATTERN.match(lines[marker_index])
-    if not marker_match:
-        return False
-
-    list_indent_stack = [
-        (
-            _indent_width(marker_match.group("indent") or ""),
-            _list_item_content_indent(marker_match),
-        )
-    ]
-
-    for idx in range(marker_index + 1, target_index + 1):
-        line = lines[idx]
-        if not line.strip(" \t\r"):
-            return False
-
-        line_indent = _indent_width(line)
-        nested_match = LIST_ITEM_PATTERN.match(line)
-        if nested_match and not _is_thematic_break_line(line):
-            marker_indent = _indent_width(nested_match.group("indent") or "")
-            while list_indent_stack and marker_indent < list_indent_stack[-1][0]:
-                list_indent_stack.pop()
-
-            if not list_indent_stack:
-                return False
-
-            if marker_indent == list_indent_stack[-1][0]:
-                list_indent_stack[-1] = (
-                    marker_indent,
-                    _list_item_content_indent(nested_match),
-                )
-                continue
-
-            if marker_indent >= list_indent_stack[-1][1]:
-                list_indent_stack.append(
-                    (
-                        marker_indent,
-                        _list_item_content_indent(nested_match),
-                    )
-                )
-                continue
-
-            return False
-
-        while len(list_indent_stack) > 1 and line_indent < list_indent_stack[-1][0]:
-            list_indent_stack.pop()
-
-        if line_indent >= list_indent_stack[-1][1]:
-            continue
-
-        return False
-
-    return True
-
-
-def _blank_run_follows_list_context(lines: list[str], blank_start: int) -> bool:
-    previous_visible_index = blank_start - 1
-    if previous_visible_index < 0 or not lines[previous_visible_index].strip(" \t\r"):
-        return False
-
-    block_start = previous_visible_index
-    while block_start > 0 and lines[block_start - 1].strip(" \t\r"):
-        block_start -= 1
-
-    for marker_index in range(previous_visible_index, block_start - 1, -1):
-        if not _starts_root_list_item(lines, marker_index):
-            continue
-        if _line_belongs_to_list_context(
-            lines, marker_index=marker_index, target_index=previous_visible_index
-        ):
-            return True
-
-    return False
-
-
-def _inject_visual_blank_line_placeholders_in_chunk(
-    text: str,
-) -> tuple[str, list[int]]:
-    """Replace internal blank lines with NBSP-only lines for Slack rendering."""
-    if not text or "\n" not in text:
-        return text, []
-
-    lines = text.split("\n")
-    rewritten: list[tuple[str, bool]] = []
-    i = 0
-
-    while i < len(lines):
-        if lines[i].strip(" \t\r"):
-            rewritten.append((lines[i], False))
-            i += 1
-            continue
-
-        blank_start = i
-        while i < len(lines) and not lines[i].strip(" \t\r"):
-            i += 1
-
-        has_visible_line_before = blank_start > 0 and bool(
-            lines[blank_start - 1].strip(" \t\r")
-        )
-        has_visible_line_after = i < len(lines) and bool(lines[i].strip(" \t\r"))
-        blank_run_follows_list_context = has_visible_line_before and (
-            _blank_run_follows_list_context(lines, blank_start)
-        )
-        next_visible_starts_reference_definition = has_visible_line_after and bool(
-            REFERENCE_DEFINITION_PATTERN.match(lines[i])
-        )
-        next_visible_starts_setext_heading = has_visible_line_after and i + 1 < len(
-            lines
-        )
-        next_visible_starts_setext_heading = bool(
-            next_visible_starts_setext_heading
-            and lines[i].strip(" \t\r")
-            and SETEXT_HEADING_UNDERLINE_PATTERN.match(lines[i + 1])
-        )
-
-        if (
-            has_visible_line_before
-            and has_visible_line_after
-            and not blank_run_follows_list_context
-            and not next_visible_starts_reference_definition
-            and not next_visible_starts_setext_heading
-        ):
-            rewritten.extend((NBSP, True) for _ in range(i - blank_start))
-        else:
-            rewritten.extend((line, False) for line in lines[blank_start:i])
-
-    rebuilt_parts: list[str] = []
-    synthetic_indices: list[int] = []
-    offset = 0
-
-    for idx, (line, is_synthetic) in enumerate(rewritten):
-        if idx > 0:
-            rebuilt_parts.append("\n")
-            offset += 1
-        if is_synthetic:
-            synthetic_indices.append(offset)
-        rebuilt_parts.append(line)
-        offset += len(line)
-
-    return "".join(rebuilt_parts), synthetic_indices
-
-
-def _inject_visual_blank_line_placeholders(text: str) -> tuple[str, list[int]]:
-    """Replace internal blank lines outside fenced code blocks."""
-    if not text or "\n" not in text:
-        return text, []
-
-    rebuilt_parts: list[str] = []
-    synthetic_indices: list[int] = []
-    offset = 0
-
-    for is_fenced, chunk in _split_fenced_code_chunks(text):
-        if is_fenced:
-            rebuilt_parts.append(chunk)
-            offset += len(chunk)
-            continue
-
-        rewritten_chunk, chunk_indices = (
-            _inject_visual_blank_line_placeholders_in_chunk(chunk)
-        )
-        rebuilt_parts.append(rewritten_chunk)
-        synthetic_indices.extend(offset + idx for idx in chunk_indices)
-        offset += len(rewritten_chunk)
-
-    return "".join(rebuilt_parts), synthetic_indices
-
-
-def _strip_synthetic_blank_line_placeholders(
-    text: str, synthetic_blank_line_indices: list[int] | None = None
-) -> str:
-    if not text or not synthetic_blank_line_indices:
-        return text
-
-    indices = set(synthetic_blank_line_indices)
-    return "".join(
-        char for idx, char in enumerate(text) if not (idx in indices and char == NBSP)
-    )
-
-
-def _remove_synthetic_space_markers(text: str) -> tuple[str, list[int]]:
-    if not text or SYNTH_SPACE_MARKER not in text:
-        return text, []
-
-    cleaned: list[str] = []
-    synthetic_indices: list[int] = []
-    mark_next_space = False
-
-    for char in text:
-        if char == SYNTH_SPACE_MARKER:
-            mark_next_space = True
-            continue
-
-        if mark_next_space and char == " ":
-            synthetic_indices.append(len(cleaned))
-        cleaned.append(char)
-        mark_next_space = False
-
-    return "".join(cleaned), synthetic_indices
-
-
-def _is_allowed_slack_angle_token(token: str) -> bool:
-    return any(pattern.match(token) for pattern in ALLOWED_SLACK_ANGLE_TOKEN_PATTERNS)
-
-
-def _find_inline_code_span_end(text: str, start: int) -> int | None:
-    """Find the end of the inline code span opened at ``start``.
-
-    Per CommonMark, a code span closes only with a backtick run of *equal*
-    length: a lone `` ` `` must not pair with the first backtick of a later
-    ``` `` ``` run. Runs of a different length are skipped whole.
-    """
-    delimiter_end = start
-    while delimiter_end < len(text) and text[delimiter_end] == "`":
-        delimiter_end += 1
-    delimiter_length = delimiter_end - start
-
-    cursor = delimiter_end
-    while True:
-        closing = text.find("`", cursor)
-        if closing == -1:
-            return None
-        run_end = closing
-        while run_end < len(text) and text[run_end] == "`":
-            run_end += 1
-        if run_end - closing == delimiter_length:
-            return run_end
-        cursor = run_end
-
-
-def _iter_inline_code_spans(text: str) -> Iterator[tuple[int, int]]:
-    """Yield ``(start, end)`` for each inline code span in ``text``.
-
-    Single source of truth for the module's span model: a span closes only on
-    a backtick run of the *same* length as its opener, may cross soft line
-    breaks, and never crosses a blank line. This matches measured Slack
-    rendering (2026-07-05): Slack pairs backticks across soft line breaks
-    within a paragraph and renders the stretch as inline code, while
-    backticks in different paragraphs never pair — so the blank-line bound
-    keeps one stray backtick from affecting anything beyond its paragraph.
-    """
-    cursor = text.find("`")
-    while cursor != -1:
-        span_end = _find_inline_code_span_end(text, cursor)
-        if span_end is None or _CODE_SPAN_BLANK_LINE_PATTERN.search(
-            text, cursor, span_end
-        ):
-            # No same-paragraph closing run: the backticks are literal text.
-            delimiter_end = cursor
-            while delimiter_end < len(text) and text[delimiter_end] == "`":
-                delimiter_end += 1
-            cursor = text.find("`", delimiter_end)
-            continue
-        yield cursor, span_end
-        cursor = text.find("`", span_end)
-
-
-def _multiline_code_span_line_map(
-    lines: list[str],
-) -> tuple[list[bool], list[bool]]:
-    """Per-line view of inline code spans that cross soft line breaks.
-
-    Returns ``(interior, continuation)``: ``interior[i]`` is True when line
-    ``i`` carries any part of a multi-line inline code span, and
-    ``continuation[i]`` is True when such a span crosses the boundary between
-    line ``i - 1`` and line ``i``. Spans are computed per non-fenced chunk —
-    backticks inside a fence are literal, and a fence delimiter terminates a
-    span (block structure binds before inline code, as in CommonMark) — using
-    the module's paragraph-bounded span model.
-    """
-    interior = [False] * len(lines)
-    continuation = [False] * len(lines)
-    chunk_indices: list[int] = []
-
-    def flush_chunk() -> None:
-        if not chunk_indices:
-            return
-        offsets: list[int] = []
-        position = 0
-        for index in chunk_indices:
-            offsets.append(position)
-            position += len(lines[index]) + 1
-        chunk_text = "\n".join(lines[index] for index in chunk_indices)
-        for span_start, span_end in _iter_inline_code_spans(chunk_text):
-            if "\n" not in chunk_text[span_start:span_end]:
-                continue
-            for local, line_start in enumerate(offsets):
-                global_index = chunk_indices[local]
-                line_end = line_start + len(lines[global_index])
-                if line_start < span_end and span_start <= line_end:
-                    interior[global_index] = True
-                    if span_start < line_start:
-                        continuation[global_index] = True
-        chunk_indices.clear()
-
-    for index, (_, is_fenced, _) in enumerate(_iter_fence_states(lines)):
-        if is_fenced:
-            flush_chunk()
-            continue
-        chunk_indices.append(index)
-    flush_chunk()
-
-    return interior, continuation
-
-
-def _transform_outside_inline_code(text: str, transform: Callable[[str], str]) -> str:
-    """Apply ``transform`` to text while keeping inline code spans verbatim.
-
-    Spans follow the module's paragraph-bounded span model
-    (``_iter_inline_code_spans``): rewriting a span's content would visibly
-    corrupt what Slack renders as code.
-
-    Spans are replaced with placeholder tokens (which contain no backticks or
-    angle brackets) rather than split out, so the transform still sees any
-    construct that *spans* a code span — e.g. an invalid angle token such as
-    ``<foo `bar` baz>`` is neutralized as a whole while the span content
-    itself stays verbatim. Reserved marker code points are stripped from the
-    input first, so crafted input cannot collide with the placeholders.
-    """
-    text = INTERNAL_MARKER_CHAR_PATTERN.sub("", text)
-
-    spans: list[str] = []
-    parts: list[str] = []
-    plain_start = 0
-
-    for start, end in _iter_inline_code_spans(text):
-        parts.append(text[plain_start:start])
-        parts.append(f"\ufff0code{len(spans)}\ufff1")
-        spans.append(text[start:end])
-        plain_start = end
-
-    parts.append(text[plain_start:])
-    transformed = transform("".join(parts))
-
-    if not spans:
-        return transformed
-    placeholder_map = {f"\ufff0code{idx}\ufff1": span for idx, span in enumerate(spans)}
-    return INLINE_CODE_PLACEHOLDER_PATTERN.sub(
-        lambda match: placeholder_map.get(match.group(0), match.group(0)),
-        transformed,
-    )
-
-
-def _transform_outside_code_regions(text: str, transform: Callable[[str], str]) -> str:
-    """Apply ``transform`` outside fenced code blocks and inline code spans.
-
-    Code samples must reach Slack verbatim: neither Slack's ``markdown`` block
-    renderer nor ``rich_text_preformatted`` interprets their content, so any
-    rewrite inside a code region is visible corruption.
-    """
-    return "".join(
-        chunk if is_fenced else _transform_outside_inline_code(chunk, transform)
-        for is_fenced, chunk in _split_fenced_code_chunks(text)
-    )
-
-
-def _is_punctuation_like(char: str, boundary_chars: set[str]) -> bool:
-    return bool(char) and char not in boundary_chars and not char.isalnum()
-
-
-def _should_preserve_raw_punctuation_emphasis(
-    source: str,
-    start: int,
-    end: int,
-    token_text: str,
-    boundary_chars: set[str],
-) -> bool:
-    tight_chars = []
-    before_char = source[start - 1] if start > 0 else ""
-    after_char = source[end] if end < len(source) else ""
-
-    if before_char and before_char not in boundary_chars:
-        tight_chars.append(before_char)
-    if after_char and after_char not in boundary_chars:
-        tight_chars.append(after_char)
-
-    if not tight_chars:
-        return False
-    if any(not _is_punctuation_like(char, boundary_chars) for char in tight_chars):
-        return False
-    # Slack only accepts ASCII punctuation (and whitespace) as a flanking
-    # neighbor. A non-ASCII punctuation neighbor — e.g. the CJK comma/period
-    # ``、``/``。`` — does not satisfy the right-/left-flanking rule, so the
-    # token must not be preserved raw; it needs the inner-ZWSP protection in
-    # ``wrap_match`` instead.
-    if any(ord(char) > 127 for char in tight_chars):
-        return False
-    if any(_is_han_or_kana_char(char) or _is_hangul_char(char) for char in token_text):
-        return False
-
-    left_idx = start - 1
-    while left_idx >= 0 and source[left_idx] in boundary_chars:
-        left_idx -= 1
-    if left_idx >= 0 and (
-        _is_han_or_kana_char(source[left_idx]) or _is_hangul_char(source[left_idx])
-    ):
-        return False
-
-    right_idx = end
-    while right_idx < len(source) and source[right_idx] in boundary_chars:
-        right_idx += 1
-    if right_idx < len(source) and (
-        _is_han_or_kana_char(source[right_idx]) or _is_hangul_char(source[right_idx])
-    ):
-        return False
-
-    return True
-
-
-def normalize_bare_urls_for_slack_markdown(text: str) -> str:
-    """Wrap bare URLs in autolink syntax for stable Slack markdown rendering."""
-    if not text:
-        return text
-
-    def wrap_chunk(chunk: str) -> str:
-        parts: list[str] = []
-        cursor = 0
-        length = len(chunk)
-
-        while cursor < length:
-            char = chunk[cursor]
-
-            if char == "<":
-                closing = chunk.find(">", cursor + 1)
-                if closing != -1:
-                    token = chunk[cursor : closing + 1]
-                    if _is_allowed_slack_angle_token(token):
-                        parts.append(token)
-                        cursor = closing + 1
-                        continue
-
-            if char == "[":
-                link_match = MARKDOWN_LINK_PATTERN.match(chunk, cursor)
-                if link_match:
-                    parts.append(link_match.group(0))
-                    cursor = link_match.end()
-                    continue
-
-            if char == "`":
-                code_span_end = _find_inline_code_span_end(chunk, cursor)
-                # Same paragraph-bounded span model as
-                # _transform_outside_inline_code: Slack pairs backticks across
-                # soft line breaks, so a URL inside such a span must stay bare
-                # — wrapping it would show a literal ``<…>`` inside the
-                # rendered code span.
-                if (
-                    code_span_end is not None
-                    and not _CODE_SPAN_BLANK_LINE_PATTERN.search(
-                        chunk, cursor, code_span_end
-                    )
-                ):
-                    parts.append(chunk[cursor:code_span_end])
-                    cursor = code_span_end
-                    continue
-                # No same-paragraph span: skip the whole backtick run, as
-                # _iter_inline_code_spans does. Restarting inside the run
-                # would open a fake shorter-delimiter span and swallow a
-                # later real span together with its content.
-                run_end = cursor
-                while run_end < len(chunk) and chunk[run_end] == "`":
-                    run_end += 1
-                parts.append(chunk[cursor:run_end])
-                cursor = run_end
-                continue
-
-            url_match = BARE_URL_PATTERN.match(chunk, cursor)
-            if url_match:
-                url = _trim_bare_url(url_match.group(0))
-                scheme = re.match(r"https?://", url, re.IGNORECASE)
-                # Only autolink when something host-like survives the trim;
-                # a bare ``https://`` followed straight by CJK would otherwise
-                # produce an empty ``<https://>`` autolink.
-                if scheme and len(url) > scheme.end():
-                    parts.append(f"<{url}>")
-                    cursor += len(url)
-                    continue
-
-            parts.append(char)
-            cursor += 1
-
-        return "".join(parts)
-
-    chunks = _split_fenced_code_chunks(text)
-    return "".join(
-        chunk if is_fenced else wrap_chunk(chunk) for is_fenced, chunk in chunks
-    )
-
-
-def sanitize_slack_text(text: str) -> str:
-    """Remove control noise and neutralize invalid Slack angle tokens.
-
-    ANSI escapes, control characters, and this module's reserved in-band
-    marker code points are removed everywhere — including code regions —
-    because they are never legitimate visible content. Angle-token
-    neutralization rewrites visible text, so it skips fenced code blocks and
-    inline code spans: a code sample containing ``<div>`` must reach Slack
-    verbatim.
-    """
-    if not text:
-        return text
-
-    cleaned = ANSI_ESCAPE_PATTERN.sub("", text)
-    cleaned = CONTROL_CHAR_PATTERN.sub("", cleaned)
-    cleaned = INTERNAL_MARKER_CHAR_PATTERN.sub("", cleaned)
-
-    def replace_invalid_token(match: re.Match[str]) -> str:
-        token = match.group(0)
-        if _is_allowed_slack_angle_token(token):
-            return token
-        return f"＜{token[1:-1]}＞"
-
-    def neutralize_angle_tokens(segment: str) -> str:
-        return SLACK_ANGLE_TOKEN_PATTERN.sub(replace_invalid_token, segment)
-
-    return _transform_outside_code_regions(cleaned, neutralize_angle_tokens)
-
-
-def _match_fence_open(line: str) -> tuple[str, int] | None:
-    match = FENCE_OPEN_PATTERN.match(line)
-    if not match:
-        return None
-    marker = match.group(1)
-    return marker[0], len(marker)
-
-
-def _is_fence_close(line: str, fence: tuple[str, int]) -> bool:
-    marker_char, marker_length = fence
-    return bool(
-        re.match(
-            rf"^[ \t]{{0,3}}{re.escape(marker_char)}{{{marker_length},}}\s*$", line
-        )
-    )
-
-
-def _iter_fence_states(lines: Iterable[str]) -> Iterator[tuple[str, bool, bool]]:
-    """Yield ``(line, is_fenced, is_opening)`` for each line.
-
-    Single source of truth for fenced-code tracking across this module.
-    ``is_fenced`` covers the fence delimiter lines themselves and the body of
-    an unclosed trailing fence; ``is_opening`` marks the opening delimiter
-    line so callers can flush per-fence state. Works with or without trailing
-    newlines on the lines.
-    """
-    active_fence: tuple[str, int] | None = None
-    for line in lines:
-        if active_fence is None:
-            opening_fence = _match_fence_open(line)
-            if opening_fence is not None:
-                active_fence = opening_fence
-                yield line, True, True
-                continue
-            yield line, False, False
-            continue
-        yield line, True, False
-        if _is_fence_close(line, active_fence):
-            active_fence = None
-
-
-def _split_fenced_code_chunks(text: str) -> list[tuple[bool, str]]:
-    chunks: list[tuple[bool, str]] = []
-    if not text:
-        return chunks
-
-    current: list[str] = []
-    current_is_fenced = False
-
-    for line, is_fenced, is_opening in _iter_fence_states(
-        text.splitlines(keepends=True)
-    ):
-        if current and (is_opening or is_fenced != current_is_fenced):
-            chunks.append((current_is_fenced, "".join(current)))
-            current = []
-        current.append(line)
-        current_is_fenced = is_fenced
-
-    if current:
-        chunks.append((current_is_fenced, "".join(current)))
-
-    return chunks
-
-
-def _normalize_underscore_emphasis_chunk(text: str) -> str:
-    # Inline code spans are paragraph-bounded (the module's span model) and
-    # must stay verbatim: Slack renders them as code, where a rewritten
-    # ``_value_`` is visible corruption. The single-line backtick alternative
-    # in PROTECTED_UNDERSCORE_SPAN_PATTERN cannot cover a span that crosses a
-    # soft line break, so the shared span walker protects those first (it
-    # also strips the reserved marker code points).
-    return _transform_outside_inline_code(text, _normalize_underscore_emphasis_prose)
-
-
-def _normalize_underscore_emphasis_prose(text: str) -> str:
-    protected_spans: list[str] = []
-
-    def protect(match: re.Match[str]) -> str:
-        token = f"\ufff0{len(protected_spans)}\ufff1"
-        protected_spans.append(match.group(0))
-        return token
-
-    normalized = PROTECTED_UNDERSCORE_SPAN_PATTERN.sub(protect, text)
-    normalized = DOUBLE_UNDERSCORE_EMPHASIS_PATTERN.sub(r"**\1**", normalized)
-    normalized = SINGLE_UNDERSCORE_EMPHASIS_PATTERN.sub(r"*\1*", normalized)
-
-    for idx, original in enumerate(protected_spans):
-        normalized = normalized.replace(f"\ufff0{idx}\ufff1", original)
-
-    return normalized
-
-
-def normalize_underscore_emphasis(text: str) -> str:
-    """Convert underscore emphasis into Slack-compatible asterisk emphasis."""
-    if not text:
-        return text
-
-    chunks = _split_fenced_code_chunks(text)
-    return "".join(
-        chunk if is_fenced else _normalize_underscore_emphasis_chunk(chunk)
-        for is_fenced, chunk in chunks
-    )
-
-
-def add_zero_width_spaces_to_markdown(text: str) -> str:
-    """Stabilize markdown rendering by padding style markers with ZWSP.
-
-    Code fences are preserved untouched.
-    """
-    formatted, _ = _format_markdown_with_spacing_metadata(text)
-    return formatted
-
-
-def _format_markdown_with_spacing_metadata(text: str) -> tuple[str, list[int]]:
-    """Return formatted markdown text plus synthetic visible-space positions."""
-    if not text:
-        return text, []
-
-    # Defense in depth for direct calls that bypass sanitize_slack_text: input
-    # carrying our reserved marker code points would collide with the inline
-    # placeholder machinery below.
-    text = INTERNAL_MARKER_CHAR_PATTERN.sub("", text)
-
-    boundary_chars = {*VISIBLE_BOUNDARY_CHARS, ZWSP, SYNTH_SPACE_MARKER}
-
-    def wrap_match(match: re.Match[str], source: str) -> str:
-        return wrap_span(source, match.start(), match.end())
-
-    def wrap_span(source: str, start: int, end: int) -> str:
-        token = source[start:end]
-        # The start/end of the chunk are effective boundaries: there is no
-        # adjacent text to separate the marker from, so they are safe. Treating
-        # them as unsafe used to append a ZWSP right after a closing marker, and
-        # when the last content character was punctuation (e.g. ``**注意:**``)
-        # the trailing ZWSP made Slack fail the CommonMark right-flanking check
-        # and exposed the literal ``**``.
-        before_safe = start == 0 or source[start - 1] in boundary_chars
-        after_safe = end == len(source) or source[end] in boundary_chars
-        if before_safe and after_safe:
-            return token
-        if _should_preserve_raw_punctuation_emphasis(
-            source, start, end, token, boundary_chars
-        ):
-            return token
-
-        # When an outer edge is tightly coupled to surrounding text, pad only
-        # that edge so Slack can treat the decoration as a standalone span.
-        # Padding a safe edge is unnecessary noise.
-        prefix = "" if before_safe else ZWSP
-        suffix = "" if after_safe else ZWSP
-
-        # Emphasis markers (``*``/``**``/``~~``) obey CommonMark delimiter-run
-        # flanking rules; inline code spans (``` `…` ```) do not. When an
-        # emphasis marker sits directly against punctuation on its inner side
-        # (``**注意:**``, ``**70%→83%**``) Slack treats the run as a delimiter
-        # only when the *outer* neighbour is whitespace or ASCII punctuation; a
-        # following CJK character or CJK punctuation (e.g. ``、``) — and even a
-        # ZWSP placed just outside the marker — leaves the literal ``**``
-        # exposed. Inserting a ZWSP just *inside* the marker makes its inner
-        # neighbour a non-punctuation character, so the run flanks via rule 2a
-        # regardless of what surrounds the token.
-        marker_char = token[0]
-        if marker_char != "`":
-            marker_len = len(token) - len(token.lstrip(marker_char))
-            open_marker = token[:marker_len]
-            inner = token[marker_len : len(token) - marker_len]
-            close_marker = token[len(token) - marker_len :]
-            inner_prefix = (
-                ZWSP if inner and _is_punctuation_like(inner[0], boundary_chars) else ""
-            )
-            inner_suffix = (
-                ZWSP
-                if inner and _is_punctuation_like(inner[-1], boundary_chars)
-                else ""
-            )
-            if inner_prefix or inner_suffix:
-                token = (
-                    f"{open_marker}{inner_prefix}{inner}{inner_suffix}{close_marker}"
-                )
-                # The inner ZWSP already lets the marker flank correctly, so an
-                # outer ZWSP on the same edge is redundant — and after a closing
-                # marker it is precisely what would re-break rendering.
-                if inner_prefix:
-                    prefix = ""
-                if inner_suffix:
-                    suffix = ""
-
-        return f"{prefix}{token}{suffix}"
-
-    def wrap_nested_code_emphasis_match(
-        match: re.Match[str],
-        source: str,
-        replacements: dict[str, dict[str, str]],
-    ) -> str:
-        start, end = match.start(), match.end()
-        before_char = source[start - 1] if start > 0 else ""
-        after_char = source[end] if end < len(source) else ""
-        strategy = _nested_code_space_strategy(source, start, end, boundary_chars)
-
-        def resolve_placeholder_raw(placeholder_match: re.Match[str]) -> str:
-            # Unknown placeholder-shaped sequences pass through unchanged
-            # (belt-and-braces against in-band collisions; the markers are
-            # already stripped at every entry point).
-            entry = replacements.get(placeholder_match.group(0))
-            return entry["raw"] if entry else placeholder_match.group(0)
-
-        resolved_text = INLINE_CODE_PLACEHOLDER_PATTERN.sub(
-            resolve_placeholder_raw, match.group(0)
-        )
-        has_ascii_word = bool(re.search(r"[A-Za-z0-9]", resolved_text))
-        adjusted_text = match.group(0)
-
-        if strategy in {"ja_zh", "ko"}:
-
-            def add_inner_spacing(placeholder_match: re.Match[str]) -> str:
-                before_inner = (
-                    adjusted_text[placeholder_match.start() - 1]
-                    if placeholder_match.start() > 0
-                    else ""
-                )
-                after_inner = (
-                    adjusted_text[placeholder_match.end()]
-                    if placeholder_match.end() < len(adjusted_text)
-                    else ""
-                )
-                prefix = (
-                    f"{SYNTH_SPACE_MARKER} "
-                    if _needs_inner_code_spacing(before_inner, boundary_chars)
-                    else ""
-                )
-                suffix = (
-                    f"{SYNTH_SPACE_MARKER} "
-                    if _needs_inner_code_spacing(after_inner, boundary_chars)
-                    else ""
-                )
-                return f"{prefix}{placeholder_match.group(0)}{suffix}"
-
-            adjusted_text = INLINE_CODE_PLACEHOLDER_PATTERN.sub(
-                add_inner_spacing, adjusted_text
-            )
-
-        if strategy == "ja_zh":
-            prefix = (
-                ""
-                if before_char in VISIBLE_BOUNDARY_CHARS or not before_char
-                else f"{SYNTH_SPACE_MARKER} "
-            )
-            suffix = (
-                ""
-                if after_char in VISIBLE_BOUNDARY_CHARS or not after_char
-                else f"{SYNTH_SPACE_MARKER} "
-            )
-            return f"{prefix}{adjusted_text}{suffix}"
-        if strategy == "ko":
-            prefix = (
-                ""
-                if before_char in VISIBLE_BOUNDARY_CHARS or not has_ascii_word
-                else f"{SYNTH_SPACE_MARKER} "
-            )
-            suffix = (
-                ""
-                if after_char in VISIBLE_BOUNDARY_CHARS or not after_char
-                else f"{SYNTH_SPACE_MARKER} "
-            )
-            return f"{prefix}{adjusted_text}{suffix}"
-        prefix = STRIP_LEFT_ZWSP_MARKER if before_char == ZWSP else ""
-        suffix = STRIP_RIGHT_ZWSP_MARKER if after_char == ZWSP else ""
-        return f"{prefix}{adjusted_text}{suffix}"
-
-    def wrap_segment(segment: str) -> tuple[str, list[int]]:
-        if not segment:
-            return segment, []
-
-        placeholder_map: dict[str, dict[str, str]] = {}
-        protected_parts: list[str] = []
-        last_end = 0
-
-        for idx, (span_start, span_end) in enumerate(_iter_inline_code_spans(segment)):
-            placeholder = f"\ufff0code{idx}\ufff1"
-            protected_parts.append(segment[last_end:span_start])
-            protected_parts.append(placeholder)
-            placeholder_map[placeholder] = {
-                "raw": segment[span_start:span_end],
-                "wrapped": wrap_span(segment, span_start, span_end),
-            }
-            last_end = span_end
-
-        protected_parts.append(segment[last_end:])
-        protected_segment = "".join(protected_parts)
-
-        placeholders_inside_emphasis: set[str] = set()
-        for pattern in EMPHASIS_PATTERNS:
-            for match in pattern.finditer(protected_segment):
-                placeholders_inside_emphasis.update(
-                    placeholder.group(0)
-                    for placeholder in INLINE_CODE_PLACEHOLDER_PATTERN.finditer(
-                        match.group(0)
-                    )
-                )
-
-        for placeholder in placeholders_inside_emphasis:
-            placeholder_map[placeholder]["wrapped"] = placeholder_map[placeholder][
-                "raw"
-            ]
-
-        for pattern in EMPHASIS_PATTERNS:
-            protected_segment = pattern.sub(
-                lambda m, s=protected_segment: (
-                    wrap_nested_code_emphasis_match(m, s, placeholder_map)
-                    if INLINE_CODE_PLACEHOLDER_PATTERN.search(m.group(0))
-                    else wrap_match(m, s)
-                ),
-                protected_segment,
-            )
-
-        def restore_placeholder(placeholder_match: re.Match[str]) -> str:
-            entry = placeholder_map.get(placeholder_match.group(0))
-            return entry["wrapped"] if entry else placeholder_match.group(0)
-
-        protected_segment = INLINE_CODE_PLACEHOLDER_PATTERN.sub(
-            restore_placeholder, protected_segment
-        )
-
-        protected_segment = re.sub(
-            f"{ZWSP}{re.escape(SYNTH_SPACE_MARKER)} ",
-            f"{SYNTH_SPACE_MARKER} ",
-            protected_segment,
-        )
-        protected_segment = re.sub(
-            f"{re.escape(SYNTH_SPACE_MARKER)} {ZWSP}",
-            f"{SYNTH_SPACE_MARKER} ",
-            protected_segment,
-        )
-        protected_segment = protected_segment.replace(
-            f"{ZWSP}{STRIP_LEFT_ZWSP_MARKER}", ""
-        )
-        protected_segment = protected_segment.replace(
-            f"{STRIP_RIGHT_ZWSP_MARKER}{ZWSP}", ""
-        )
-        protected_segment = protected_segment.replace(STRIP_LEFT_ZWSP_MARKER, "")
-        protected_segment = protected_segment.replace(STRIP_RIGHT_ZWSP_MARKER, "")
-        protected_segment = re.sub(f"{ZWSP}+", ZWSP, protected_segment)
-        return _remove_synthetic_space_markers(protected_segment)
-
-    chunks = _split_fenced_code_chunks(text)
-    combined_parts: list[str] = []
-    combined_indices: list[int] = []
-    offset = 0
-    for is_fenced, chunk in chunks:
-        if is_fenced:
-            combined_parts.append(chunk)
-            offset += len(chunk)
-            continue
-        formatted_chunk, synthetic_indices = wrap_segment(chunk)
-        combined_parts.append(formatted_chunk)
-        combined_indices.extend(offset + idx for idx in synthetic_indices)
-        offset += len(formatted_chunk)
-
-    return "".join(combined_parts), combined_indices
-
-
-# Backward-compatible alias
-add_zero_width_spaces = add_zero_width_spaces_to_markdown
-
-
-def _match_slack_angle_token_end(text: str, start: int) -> int | None:
-    """Return the end index (exclusive) of a valid Slack angle token at ``start``.
-
-    Only recognized Slack tokens (links, mentions, ``<!date^…>``) protect
-    their inner pipes from cell splitting. A bare ``<`` — e.g. the comparison
-    in ``x < y`` — is literal text and must not open a protected region;
-    a stateful "inside angle brackets" flag did exactly that, so one lone
-    ``<`` swallowed every later pipe on the line into a single cell.
-    """
-    if start >= len(text) or text[start] != "<":
-        return None
-    closing = text.find(">", start + 1)
-    if closing == -1:
-        return None
-    token = text[start : closing + 1]
-    return closing + 1 if _is_allowed_slack_angle_token(token) else None
-
-
-def _split_markdown_table_cells(line: str) -> list[str]:
-    """Split markdown table cells while preserving pipes inside Slack tokens."""
-    working = line.strip()
-    if not working:
-        return []
-
-    if working.startswith("|"):
-        working = working[1:]
-    if working.endswith("|"):
-        working = working[:-1]
-
-    cells: list[str] = []
-    current: list[str] = []
-    escaped = False
-    cursor = 0
-
-    while cursor < len(working):
-        ch = working[cursor]
-        if escaped:
-            current.append(ch)
-            escaped = False
-            cursor += 1
-            continue
-
-        if ch == "\\":
-            current.append(ch)
-            escaped = True
-            cursor += 1
-            continue
-
-        if ch == "`":
-            code_span_end = _find_inline_code_span_end(working, cursor)
-            if code_span_end is not None:
-                current.append(working[cursor:code_span_end])
-                cursor = code_span_end
-                continue
-
-        if ch == "<":
-            token_end = _match_slack_angle_token_end(working, cursor)
-            if token_end is not None:
-                current.append(working[cursor:token_end])
-                cursor = token_end
-                continue
-
-        if ch == "|":
-            cells.append("".join(current).strip())
-            current = []
-            cursor += 1
-            continue
-
-        current.append(ch)
-        cursor += 1
-
-    cells.append("".join(current).strip())
-    return cells
-
-
-def _count_cell_words(cell_text: str) -> int:
-    tokens = [token for token in cell_text.strip().split() if token]
-    return len(tokens) or 1
-
-
-def _split_heading_prefix_and_first_cell(
-    heading_prefix: str, reference_cell: str | None
-) -> tuple[str, str] | None:
-    tokens = [token for token in heading_prefix.strip().split() if token]
-    if len(tokens) < 2:
-        return None
-
-    reference_words = _count_cell_words(reference_cell or "")
-    first_cell_words = min(reference_words, len(tokens) - 1)
-    if first_cell_words <= 0:
-        return None
-    if first_cell_words != reference_words:
-        # The heading tail cannot supply a first cell shaped like the
-        # reference row's first cell, so this is a heading that merely
-        # contains a pipe (``## Phase 1 | Overview`` followed by prose with
-        # a pipe), not a glued table header.
-        return None
-
-    heading_tokens = tokens[:-first_cell_words]
-    first_cell_tokens = tokens[-first_cell_words:]
-    if not heading_tokens or not first_cell_tokens:
-        return None
-
-    return " ".join(heading_tokens), " ".join(first_cell_tokens)
-
-
-def _split_heading_and_table_row(
-    line: str, next_line: str | None = None
-) -> tuple[str, str] | None:
-    """Split lines like '# Heading |a|b|' into heading and table row.
-
-    The split targets one specific LLM failure mode: a table's header row
-    glued onto the heading line, with the real data rows following. A heading
-    that merely contains a pipe (``## Phase 1 | Overview``) is not that
-    pattern, so the split requires a table-like (pipe-carrying) next line;
-    without one the heading is left intact.
-    """
-    if "|" not in line:
-        return None
-    if not next_line or "|" not in next_line:
-        return None
-
-    escaped = False
-    first_pipe = -1
-    cursor = 0
-    while cursor < len(line):
-        ch = line[cursor]
-        if escaped:
-            escaped = False
-            cursor += 1
-            continue
-        if ch == "\\":
-            escaped = True
-            cursor += 1
-            continue
-        if ch == "`":
-            code_span_end = _find_inline_code_span_end(line, cursor)
-            if code_span_end is not None:
-                cursor = code_span_end
-                continue
-        if ch == "<":
-            token_end = _match_slack_angle_token_end(line, cursor)
-            if token_end is not None:
-                cursor = token_end
-                continue
-        if ch == "|":
-            first_pipe = cursor
-            break
-        cursor += 1
-
-    if first_pipe < 0:
-        return None
-
-    heading_part = line[:first_pipe].rstrip()
-    table_part = line[first_pipe:].strip()
-
-    if not heading_part or not heading_part.lstrip().startswith("#"):
-        return None
-
-    heading_match = re.match(r"^([ \t]{0,3}#{1,6})\s+(.+)$", heading_part)
-    if not heading_match:
-        return None
-
-    heading_marker = heading_match.group(1)
-    heading_body = heading_match.group(2).strip()
-
-    if not table_part.startswith("|"):
-        table_part = "|" + table_part
-    if not table_part.endswith("|"):
-        table_part = table_part + "|"
-
-    explicit_cells = _split_markdown_table_cells(table_part)
-    if not explicit_cells:
-        return None
-
-    reference_cell: str | None = None
-    if (
-        next_line
-        and "|" in next_line
-        and not LOOSE_TABLE_SEPARATOR_PATTERN.match(next_line.strip())
-    ):
-        next_working = next_line.strip()
-        if not next_working.startswith("|"):
-            next_working = "|" + next_working
-        if not next_working.endswith("|"):
-            next_working = next_working + "|"
-        next_cells = _split_markdown_table_cells(next_working)
-        if next_cells:
-            reference_cell = next_cells[0]
-
-    if reference_cell is None and explicit_cells:
-        reference_cell = explicit_cells[0]
-
-    split_heading = _split_heading_prefix_and_first_cell(heading_body, reference_cell)
-    if not split_heading:
-        return None
-
-    heading_text_body, first_cell = split_heading
-    heading_text = f"{heading_marker} {heading_text_body}"
-    table_cells = [first_cell] + explicit_cells
-    table_line = "|" + "|".join(table_cells) + "|"
-
-    return heading_text, table_line
-
-
-def normalize_markdown_tables(markdown_text: str) -> str:
-    """Normalize markdown tables: pipe completion, separator completion, column sizing."""
-    if not markdown_text:
-        return markdown_text
-
-    lines = markdown_text.splitlines()
-    normalized: list[str] = []
-    buffer: list[str] = []
-
-    def is_table_block(candidates: list[str]) -> bool:
-        if len(candidates) < 2:
-            return False
-        if any(
-            LOOSE_TABLE_SEPARATOR_PATTERN.match(line.strip()) for line in candidates
-        ):
-            return True
-
-        column_counts: list[int] = []
-        for line in candidates:
-            working = line.strip()
-            if "|" not in working:
-                continue
-            if not working.startswith("|"):
-                working = "|" + working
-            if not working.endswith("|"):
-                working = working + "|"
-            if TABLE_SEPARATOR_PATTERN.match(working):
-                continue
-            column_counts.append(len(_split_markdown_table_cells(working)))
-
-        if len(column_counts) < 2:
-            return False
-
-        min_cols, max_cols = min(column_counts), max(column_counts)
-        if max_cols < 2:
-            return False
-
-        return max_cols - min_cols <= 1
-
-    def flush_buffer() -> None:
-        nonlocal buffer
-        if buffer and is_table_block(buffer):
-            header_line = buffer[0].strip()
-            if not header_line.startswith("|"):
-                header_line = "|" + header_line
-            if not header_line.endswith("|"):
-                header_line = header_line + "|"
-            header_cells = _split_markdown_table_cells(header_line)
-            column_count = max(1, len(header_cells))
-
-            def normalize_row(raw_line: str) -> str:
-                if LOOSE_TABLE_SEPARATOR_PATTERN.match(raw_line.strip()):
-                    return ""
-                working = raw_line.strip()
-                if "|" not in working:
-                    return ""
-                if not working.startswith("|"):
-                    working = "|" + working
-                if not working.endswith("|"):
-                    working = working + "|"
-                cells = _split_markdown_table_cells(working)
-                cells = (cells + [""] * column_count)[:column_count]
-                return "|" + "|".join(cells) + "|"
-
-            normalized.append(normalize_row(buffer[0]))
-            normalized.append("|" + "|".join(["---"] * column_count) + "|")
-            for line in buffer[1:]:
-                row = normalize_row(line)
-                if row:
-                    normalized.append(row)
-        else:
-            normalized.extend(buffer)
-        buffer = []
-
-    span_interior, _ = _multiline_code_span_line_map(lines)
-
-    for idx, (line, is_fenced, is_opening) in enumerate(_iter_fence_states(lines)):
-        if is_fenced:
-            if is_opening:
-                flush_buffer()
-            normalized.append(line)
-            continue
-
-        if span_interior[idx]:
-            # The line carries part of a multi-line inline code span: Slack
-            # renders that stretch as code, so it is never a table row or a
-            # glued heading, however pipe-like it looks.
-            flush_buffer()
-            normalized.append(line)
-            continue
-
-        stripped = line.strip()
-
-        next_line = lines[idx + 1] if idx + 1 < len(lines) else None
-        if next_line is not None and span_interior[idx + 1]:
-            # A line carrying part of a multi-line code span is code, not
-            # table evidence: it must neither justify a glued-heading split
-            # nor count as the separator behind an ATX-looking header row.
-            next_line = None
-        heading_and_table = _split_heading_and_table_row(line, next_line)
-        if heading_and_table:
-            flush_buffer()
-            heading_text, table_line = heading_and_table
-            normalized.append(heading_text)
-            buffer.append(table_line)
-        elif "|" in stripped and (
-            buffer
-            or not ATX_HEADING_PATTERN.match(line)
-            or bool(
-                next_line is not None
-                and LOOSE_TABLE_SEPARATOR_PATTERN.match(next_line.strip())
-            )
-        ):
-            # A heading-looking line whose glued-header split was rejected
-            # only escapes buffering when it would *start* a candidate run: a
-            # heading that merely contains a pipe must not seed the
-            # pipe-completion heuristic. Inside an already-buffered run it is
-            # kept — there it is a data row whose first cell begins with '#'
-            # — and a following separator row proves a table context, so a
-            # header row whose first cell begins with '#' seeds the run too.
-            buffer.append(line)
-        else:
-            flush_buffer()
-            normalized.append(line)
-    flush_buffer()
-
-    return "\n".join(normalized)
-
-
-def _looks_like_markdown_table(text: str) -> bool:
-    """Heuristic check for markdown table candidates."""
-    lines = [line for line in text.splitlines() if line.strip()]
-    if len(lines) < 2:
-        return False
-    table_like_lines = sum(
-        1
-        for line in lines
-        if line.strip().startswith("|") and line.strip().endswith("|")
-    )
-    return table_like_lines >= 2
-
-
-def _slack_mention_element(mention: str) -> dict[str, Any]:
-    """Map a Slack mention token body to its rich_text element.
-
-    ``mention`` is the token interior without the angle brackets or ``|label``
-    (e.g. ``@U123``, ``#C123``, ``!subteam^S123``, ``!here``). Slack renders
-    these from the id alone, so no display text is carried.
-    """
-    sigil, body = mention[0], mention[1:]
-    if sigil == "@":
-        return {"type": "user", "user_id": body}
-    if sigil == "#":
-        return {"type": "channel", "channel_id": body}
-    if body.startswith("subteam^"):
-        return {"type": "usergroup", "usergroup_id": body[len("subteam^") :]}
-    return {"type": "broadcast", "range": body}
-
-
-def _slack_mention_element_to_token(element: dict[str, Any]) -> str:
-    """Inverse of :func:`_slack_mention_element` for plain-text fallbacks.
-
-    Emitting the canonical ``<#C…>`` / ``<@U…>`` token (rather than an empty
-    string) keeps the mention live when a rich_text block is downgraded to a
-    mrkdwn fallback, so it still links and notifies.
-    """
-    element_type = element.get("type")
-    if element_type == "user":
-        return f"<@{element.get('user_id', '')}>"
-    if element_type == "channel":
-        return f"<#{element.get('channel_id', '')}>"
-    if element_type == "usergroup":
-        return f"<!subteam^{element.get('usergroup_id', '')}>"
-    return f"<!{element.get('range', '')}>"
-
-
-def _create_rich_text_inline_elements(
-    text: str, *, empty_text: str = ""
-) -> list[dict[str, Any]]:
-    """Build Slack rich text inline elements from a small markdown fragment."""
-    clean_text = strip_zero_width_spaces(text or "")
-    clean_text = clean_text.replace("\\|", "|")
-    if not clean_text.strip():
-        clean_text = empty_text
-
-    elements: list[dict[str, Any]] = []
-    last_index = 0
-
-    for match in TABLE_TOKEN_PATTERN.finditer(clean_text):
-        if match.start() > last_index:
-            prefix = clean_text[last_index : match.start()]
-            if prefix:
-                elements.append({"type": "text", "text": prefix})
-
-        element: dict[str, Any]
-        markdown_label = match.group("markdown_label")
-        markdown_url = match.group("markdown_url")
-        angle_url = match.group("angle_url")
-        angle_label = match.group("angle_label")
-        mention = match.group("mention")
-        token = match.group("token") or ""
-
-        if markdown_label and markdown_url:
-            element = {"type": "link", "url": markdown_url, "text": markdown_label}
-        elif angle_url:
-            element = {
-                "type": "link",
-                "url": angle_url,
-                "text": angle_label or angle_url,
-            }
-        elif mention:
-            element = _slack_mention_element(mention)
-        else:
-            style: dict[str, bool] = {}
-            content = token
-
-            if content.startswith("`") and content.endswith("`"):
-                delimiter_len = len(match.group("code_delimiter") or "`")
-                content = content[delimiter_len:-delimiter_len]
-                style["code"] = True
-            elif content.startswith("~~") and content.endswith("~~"):
-                content = content[2:-2]
-                style["strike"] = True
-            elif content.startswith("**") and content.endswith("**"):
-                content = content[2:-2]
-                style["bold"] = True
-            elif content.startswith("*") and content.endswith("*"):
-                content = content[1:-1]
-                style["italic"] = True
-
-            # A link wrapped entirely in emphasis (``**[text](url)**``) is matched
-            # by the emphasis branch above, not the link branch, so its inner
-            # content is a bare ``[text](url)``. Emit a styled ``link`` element
-            # rather than a literal text run, otherwise the link is dead in Slack.
-            inner_link = (
-                TABLE_TOKEN_PATTERN.fullmatch(content)
-                if style and not style.get("code")
-                else None
-            )
-            if inner_link is not None and inner_link.group("markdown_url"):
-                element = {
-                    "type": "link",
-                    "url": inner_link.group("markdown_url"),
-                    "text": inner_link.group("markdown_label"),
-                    "style": style,
-                }
-            else:
-                element = {"type": "text", "text": content}
-                if style:
-                    element["style"] = style
-        elements.append(element)
-        last_index = match.end()
-
-    if last_index < len(clean_text):
-        suffix = clean_text[last_index:]
-        elements.append({"type": "text", "text": suffix})
-
-    if not elements:
-        elements.append({"type": "text", "text": clean_text})
-
-    return elements
-
-
-def _create_rich_text_section(text: str, *, empty_text: str = "") -> dict[str, Any]:
-    return {
-        "type": "rich_text_section",
-        "elements": _create_rich_text_inline_elements(text, empty_text=empty_text),
-    }
-
-
-def _create_table_cell(text: str) -> dict[str, Any]:
-    """Build Slack rich_text cell from markdown fragment."""
-    return {
-        "type": "rich_text",
-        "elements": [_create_rich_text_section(text, empty_text="-")],
-    }
-
-
-def extract_plain_text_from_table_cell(cell: dict[str, Any]) -> str:
-    """Extract plain text from a Slack table cell object."""
-    if not isinstance(cell, dict):
-        return ""
-
-    if cell.get("type") == "rich_text":
-        texts: list[str] = []
-        for element in cell.get("elements", []):
-            if not isinstance(element, dict):
-                continue
-            if element.get("type") == "rich_text_section":
-                texts.append(
-                    _rich_text_inline_elements_to_plain_text(
-                        element.get("elements", [])
-                    )
-                )
-            elif "text" in element:
-                texts.append(str(element.get("text", "")))
-        return "".join(texts)
-
-    return str(cell.get("text", ""))
-
-
-def markdown_table_to_slack_table(table_markdown: str) -> dict[str, Any] | None:
-    """Convert markdown table text to Slack table block."""
-    lines = [
-        line.rstrip() for line in table_markdown.strip().splitlines() if line.strip()
-    ]
-    rows: list[list[dict[str, Any]]] = []
-    expected_columns: int | None = None
-
-    for line in lines:
-        if TABLE_SEPARATOR_PATTERN.match(line):
-            continue
-        if "|" not in line:
-            continue
-
-        cells = _split_markdown_table_cells(line)
-        if not cells:
-            continue
-
-        if expected_columns is None:
-            expected_columns = max(1, len(cells))
-        else:
-            cells = (cells + [""] * expected_columns)[:expected_columns]
-
-        rows.append(
-            [_create_table_cell(cell if cell.strip() else "-") for cell in cells]
-        )
-
-    if not rows:
-        return None
-
-    return {"type": "table", "rows": rows}
-
-
-# Backward-compatible alias
-markdown_table_to_table_block = markdown_table_to_slack_table
-
-
-def _truncate_plain_text(text: str, max_length: int) -> str:
-    if len(text) <= max_length:
-        return text
-    return text[: max_length - 1].rstrip() + "…"
-
-
-def _is_http_url(url: str) -> bool:
-    try:
-        parsed = urlparse(url)
-    except ValueError:
-        return False
-    return parsed.scheme in {"http", "https"} and bool(parsed.netloc)
-
-
-def _rich_text_inline_elements_to_plain_text(elements: list[dict[str, Any]]) -> str:
-    texts: list[str] = []
-    for element in elements or []:
-        if not isinstance(element, dict):
-            continue
-        element_type = element.get("type")
-        if element_type == "link":
-            texts.append(str(element.get("text") or element.get("url", "")))
-        elif element_type in {"user", "channel", "usergroup", "broadcast"}:
-            texts.append(_slack_mention_element_to_token(element))
-        else:
-            texts.append(str(element.get("text", "")))
-    return "".join(texts)
-
-
-def _rich_text_object_to_plain_text(element: dict[str, Any]) -> str:
-    element_type = element.get("type")
-    if element_type == "rich_text_section":
-        return _rich_text_inline_elements_to_plain_text(element.get("elements", []))
-    if element_type in {"rich_text_preformatted", "rich_text_quote"}:
-        return _rich_text_inline_elements_to_plain_text(element.get("elements", []))
-    if element_type == "rich_text_list":
-        style = element.get("style")
-        indent = max(0, int(element.get("indent") or 0))
-        offset = max(0, int(element.get("offset") or 0))
-        prefix = "  " * indent
-        lines: list[str] = []
-        for idx, child in enumerate(element.get("elements", []), start=1):
-            if not isinstance(child, dict):
-                continue
-            child_text = _rich_text_object_to_plain_text(child).strip()
-            if not child_text:
-                continue
-            marker = f"{offset + idx}." if style == "ordered" else "-"
-            lines.append(f"{prefix}{marker} {child_text}")
-        return "\n".join(lines)
-    return ""
-
-
-def _rich_text_block_to_plain_text(block: dict[str, Any]) -> str:
-    annotated_plain_text = getattr(block, "_plain_text", None)
-    if annotated_plain_text:
-        return str(annotated_plain_text).strip()
-
-    parts = [
-        _rich_text_object_to_plain_text(element)
-        for element in block.get("elements", [])
-        if isinstance(element, dict)
-    ]
-    return "\n".join(part for part in parts if part).strip()
-
-
-def _plain_text_from_markdown_fragment(text: str) -> str:
-    return _rich_text_inline_elements_to_plain_text(
-        _create_rich_text_inline_elements(text)
-    ).strip()
-
-
-def _create_markdown_block(
-    content: str, *, preserve_visual_blank_lines: bool = False
-) -> dict[str, Any] | None:
-    formatted, synthetic_indices = _format_markdown_with_spacing_metadata(content)
-    plain_text = _build_markdown_block_plain_text(formatted, synthetic_indices)
-    synthetic_blank_line_indices: list[int] = []
-    if preserve_visual_blank_lines:
-        formatted, synthetic_blank_line_indices = (
-            _inject_visual_blank_line_placeholders(formatted)
-        )
-    if not formatted.strip():
-        return None
-
-    block = _AnnotatedSlackBlock({"type": "markdown", "text": formatted})
-    block._plain_text = plain_text
-    block._synthetic_space_indices = synthetic_indices
-    block._synthetic_blank_line_indices = synthetic_blank_line_indices
-    block._expansion_items = _estimate_markdown_expansion_items(formatted)
-    return block
-
-
-def _is_markdown_expansion_breaker_line(line: str) -> bool:
-    """Return True for lines Slack expands into their own top-level block.
-
-    ATX headings and thematic breaks each become one expansion item and end
-    the surrounding content run. A setext ``===`` underline is counted too;
-    that can over-count by one (heading text + underline), which only makes
-    the estimate conservative.
-    """
-    if ATX_HEADING_PATTERN.match(line):
-        return True
-    if _is_thematic_break_line(line):
-        return True
-    stripped = line.strip()
-    return bool(stripped) and set(stripped) == {"="}
-
-
-def _estimate_markdown_expansion_items(text: str) -> int:
-    """Estimate how many native blocks Slack expands this markdown text into.
-
-    Model measured against a real workspace (see the constants above): each
-    heading / thematic break is one item, and each maximal run of any other
-    content *between* those breakers is one item — blank lines inside a run
-    do not split it. Fenced lines always count as run content.
-    """
-    items = 0
-    in_run = False
-    for line, is_fenced, _ in _iter_fence_states(text.split("\n")):
-        if is_fenced:
-            if not in_run:
-                items += 1
-                in_run = True
-            continue
-        if not line.strip():
-            continue
-        if _is_markdown_expansion_breaker_line(line):
-            items += 1
-            in_run = False
-            continue
-        if not in_run:
-            items += 1
-            in_run = True
-    return max(1, items)
-
-
-def _split_text_at_blank_lines(text: str, max_length: int, max_items: int) -> list[str]:
-    """Greedily pack paragraph units into pieces within both budgets.
-
-    Units are separated by blank-line runs outside fenced code (blank lines
-    inside a fence never split). The blank run at a piece boundary is dropped
-    — adjacent Slack blocks already render separated — while blank runs
-    packed inside a piece are kept verbatim. A piece is closed when adding
-    the next unit would exceed ``max_length`` characters or ``max_items``
-    estimated expansion items. A single unit over either budget is returned
-    oversized; the caller splits it harder.
-    """
-    if (
-        len(text) <= max_length
-        and _estimate_markdown_expansion_items(text) <= max_items
-    ):
-        return [text]
-
-    units: list[list[str]] = []
-    content: list[str] = []
-    blanks: list[str] = []
-    for line, is_fenced, _ in _iter_fence_states(text.split("\n")):
-        if not is_fenced and not line.strip():
-            if content:
-                blanks.append(line)
-            else:
-                # Leading blank lines stay attached to the first unit.
-                content.append(line)
-            continue
-        if blanks:
-            units.append(content)
-            units.append(blanks)
-            content, blanks = [], []
-        content.append(line)
-    if content:
-        units.append(content)
-    if blanks:
-        units.append(blanks)
-
-    pieces: list[str] = []
-    current: list[str] = []
-    pending_blanks: list[str] = []
-    for index in range(0, len(units), 2):
-        unit = units[index]
-        candidate = current + (pending_blanks if current else []) + unit
-        candidate_text = "\n".join(candidate)
-        if current and (
-            len(candidate_text) > max_length
-            or _estimate_markdown_expansion_items(candidate_text) > max_items
-        ):
-            pieces.append("\n".join(current))
-            current = list(unit)
-        else:
-            current = candidate
-        pending_blanks = units[index + 1] if index + 1 < len(units) else []
-    if current:
-        pieces.append("\n".join(current))
-    return pieces
-
-
-_QUOTE_MARKER_PREFIX_PATTERN = re.compile(r"^[ \t]{0,3}>[ \t]?")
-
-
-def _split_line_verbatim_to_length(line: str, max_length: int) -> list[str]:
-    """Hard-cut one overlong line into parts, preserving every character.
-
-    For fenced code lines only: the prose splitter prefers a space boundary
-    and drops that space — silently altering code — and its quote/list
-    marker handling must not apply to a code line that merely starts with
-    ``> `` or ``- ``.
-    """
-    return [line[i : i + max_length] for i in range(0, len(line), max_length)]
-
-
-def _split_single_line_to_length(line: str, max_length: int) -> list[str]:
-    """Split one overlong prose line, preferring a space boundary near the
-    limit.
-
-    A leading quote or list marker is never split off on its own (the space
-    search starts after it), and a quote marker is re-applied to continuation
-    parts so consecutive parts keep rendering as one quote. List continuations
-    stay unmarked: as following lines they are lazy item continuations.
-    """
-    quote_match = _QUOTE_MARKER_PREFIX_PATTERN.match(line)
-    marker_match = quote_match or LIST_ITEM_PATTERN.match(line)
-    prefix_end = marker_match.end() if marker_match else 0
-    continuation_prefix = quote_match.group(0) if quote_match else ""
-
-    parts: list[str] = []
-    while len(line) > max_length:
-        cut = line.rfind(" ", prefix_end + 1, max_length + 1)
-        if cut <= prefix_end:
-            parts.append(line[:max_length])
-            line = continuation_prefix + line[max_length:]
-        else:
-            parts.append(line[:cut])
-            line = continuation_prefix + line[cut + 1 :]
-        prefix_end = len(continuation_prefix)
-    parts.append(line)
-    return parts
-
-
-def _split_lines_to_length(text: str, max_length: int, max_items: int) -> list[str]:
-    """Split at line boundaries into pieces within both budgets.
-
-    Last-resort splitter for content that exceeds a budget without blank-line
-    split points (a single huge paragraph, a fence body, or a long run of
-    headings). When the cut lands inside an (unclosed) fence, the continuation
-    piece re-opens the fence with the original delimiter line so both pieces
-    keep rendering as code.
-    """
-    pieces: list[str] = []
-    current: list[str] = []
-    current_len = 0
-    current_items = 0
-    in_run = False
-    active_fence_open: str | None = None
-    lines = text.split("\n")
-    _, span_continuation = _multiline_code_span_line_map(lines)
-    # Keeping a span-crossing boundary glued may exceed the packing target;
-    # the valve bounds that overshoot so shrink-and-retry still converges and
-    # the hard block limit keeps real headroom. A span larger than the valve
-    # is cut anyway (documented limitation).
-    span_valve = min(max_length + 512, SLACK_MAX_MARKDOWN_TEXT_LENGTH - 256)
-    # Total length of the span-connected line group starting at each line,
-    # used to flush *before* a span opener when the accumulated piece would
-    # otherwise push the span past the valve even though the span alone fits.
-    span_group_total = [0] * len(lines)
-    for index in range(len(lines) - 1, -1, -1):
-        continues = index + 1 < len(lines) and span_continuation[index + 1]
-        span_group_total[index] = len(lines[index]) + (
-            1 + span_group_total[index + 1] if continues else 0
-        )
-
-    def flush(next_fence_prefix: str | None) -> None:
-        nonlocal current, current_len, current_items, in_run
-        # A piece holding nothing but the fence-open line being reopened is a
-        # synthetic duplicate (the opener was consumed, the first body part
-        # overflowed): emitting it would post a stray delimiter-only block.
-        if current and not (
-            next_fence_prefix is not None and current == [next_fence_prefix]
-        ):
-            pieces.append("\n".join(current))
-        if next_fence_prefix:
-            current = [next_fence_prefix]
-            current_len = len(next_fence_prefix)
-            current_items = 1
-            in_run = True
-        else:
-            current = []
-            current_len = 0
-            current_items = 0
-            in_run = False
-
-    for line_index, (line, is_fenced, is_opening) in enumerate(
-        _iter_fence_states(lines)
-    ):
-        if is_opening:
-            active_fence_open = line
-        elif not is_fenced:
-            active_fence_open = None
-
-        # A fence-close line never starts a new piece: flushing on it would
-        # emit the close (plus reopened fence) as a stray delimiter-only
-        # block. Appending it may exceed the packing target by a few
-        # characters, well within the headroom to the hard limit.
-        line_is_fence_close = False
-        if is_fenced and not is_opening and active_fence_open is not None:
-            open_spec = _match_fence_open(active_fence_open)
-            line_is_fence_close = open_spec is not None and _is_fence_close(
-                line, open_spec
-            )
-
-        line_is_blank = not is_fenced and not line.strip()
-        line_is_breaker = (
-            not is_fenced
-            and not line_is_blank
-            and _is_markdown_expansion_breaker_line(line)
-        )
-
-        for part_index, part in enumerate(
-            [line]
-            if len(line) <= max_length
-            else (
-                _split_line_verbatim_to_length(line, max_length)
-                if is_fenced
-                else _split_single_line_to_length(line, max_length)
-            )
-        ):
-            # Word-split continuations of a breaker line render as plain
-            # content, so only the first part keeps the breaker class.
-            is_breaker = line_is_breaker and part_index == 0
-            if line_is_blank:
-                part_items = 0
-            elif is_breaker:
-                part_items = 1
-            else:
-                part_items = 0 if in_run else 1
-
-            added = len(part) + (1 if current else 0)
-            keep_span_together = (
-                part_index == 0
-                and span_continuation[line_index]
-                and current_len + added <= span_valve
-            )
-            opens_fitting_span = (
-                part_index == 0
-                and len(line) <= max_length
-                and not span_continuation[line_index]
-                and line_index + 1 < len(lines)
-                and span_continuation[line_index + 1]
-                and span_group_total[line_index] <= span_valve
-            )
-            span_needs_fresh_piece = (
-                opens_fitting_span
-                and current_len + (1 if current else 0) + span_group_total[line_index]
-                > span_valve
-            )
-            if (
-                current
-                and not line_is_fence_close
-                and not keep_span_together
-                and (
-                    current_len + added > max_length
-                    or current_items + part_items > max_items
-                    or span_needs_fresh_piece
-                )
-            ):
-                reopen = active_fence_open if active_fence_open != part else None
-                flush(reopen)
-                added = len(part) + (1 if current else 0)
-                if line_is_blank:
-                    part_items = 0
-                elif is_breaker:
-                    part_items = 1
-                else:
-                    part_items = 0 if in_run else 1
-            current.append(part)
-            current_len += added
-            current_items += part_items
-            if is_breaker:
-                in_run = False
-            elif not line_is_blank:
-                in_run = True
-
-    if current:
-        pieces.append("\n".join(current))
-    return pieces
-
-
-def _markdown_block_fits_slack_limits(block: dict[str, Any]) -> bool:
-    return (
-        len(block["text"]) <= SLACK_MAX_MARKDOWN_TEXT_LENGTH
-        and _estimate_markdown_expansion_items(block["text"])
-        <= SLACK_MAX_EXPANSION_ITEMS_PER_MESSAGE
-    )
-
-
-def _create_markdown_blocks(
-    content: str, *, preserve_visual_blank_lines: bool = False
-) -> list[dict[str, Any]]:
-    """Build ``markdown`` blocks that fit Slack's measured hard limits.
-
-    The whole content is tried as a single block first. Only when the
-    *formatted* text exceeds ``SLACK_MAX_MARKDOWN_TEXT_LENGTH`` or the
-    estimated server-side expansion exceeds the per-message item budget is
-    the raw content split — at paragraph boundaries when possible, then at
-    line/word boundaries — and each piece re-checked after formatting,
-    shrinking the packing budget geometrically until every block fits.
-    """
-
-    def build(piece: str) -> dict[str, Any] | None:
-        return _create_markdown_block(
-            piece, preserve_visual_blank_lines=preserve_visual_blank_lines
-        )
-
-    whole = build(content)
-    if whole is None:
-        return []
-    if _markdown_block_fits_slack_limits(whole):
-        return [whole]
-
-    blocks: list[dict[str, Any]] = []
-    worklist: list[tuple[str, int, int]] = [
-        (content, _MARKDOWN_SPLIT_TARGET_LENGTH, _MARKDOWN_EXPANSION_ITEMS_TARGET)
-    ]
-    while worklist:
-        piece, budget, items_budget = worklist.pop(0)
-        block = build(piece)
-        if block is None:
-            continue
-        if _markdown_block_fits_slack_limits(block):
-            blocks.append(block)
-            continue
-
-        sub_pieces = _split_text_at_blank_lines(piece, budget, items_budget)
-        if len(sub_pieces) == 1:
-            sub_pieces = _split_lines_to_length(piece, budget, items_budget)
-        if len(sub_pieces) > 1:
-            worklist = [(sub, budget, items_budget) for sub in sub_pieces] + worklist
-            continue
-        if budget > 256 or items_budget > 8:
-            # The piece fits the raw budgets but its *formatted* text overflows
-            # (ZWSP/NBSP inflation or estimation drift): shrink and retry.
-            worklist.insert(
-                0,
-                (
-                    piece,
-                    max(256, int(budget * 0.8)),
-                    max(8, int(items_budget * 0.8)),
-                ),
-            )
-            continue
-        # A floor-budget piece cannot exceed the hard limits, so this is
-        # unreachable; keep the block rather than loop forever.
-        blocks.append(block)
-    return blocks
-
-
-def _create_rich_text_block(
-    elements: list[dict[str, Any]], *, plain_text: str | None = None
-) -> dict[str, Any]:
-    block = _AnnotatedSlackBlock({"type": "rich_text", "elements": elements})
-    if plain_text is not None:
-        block._plain_text = plain_text
-    return block
-
-
-def _create_image_block_from_line(line: str) -> dict[str, Any] | None:
-    match = STANDALONE_IMAGE_PATTERN.match(line)
-    if not match:
-        return None
-
-    image_url = match.group("url").strip()
-    if not _is_http_url(image_url) or len(image_url) > 3000:
-        return None
-
-    alt_text = _plain_text_from_markdown_fragment(match.group("alt") or "")
-    alt_text = _truncate_plain_text(alt_text or "Image", 2000)
-    return {"type": "image", "image_url": image_url, "alt_text": alt_text}
-
-
-def _is_setext_heading_underline(lines: list[str], index: int) -> bool:
-    if index <= 0:
-        return False
-    line = lines[index].strip()
-    if not line or set(line) != {"-"}:
-        return False
-    return bool(lines[index - 1].strip())
-
-
-def _create_divider_block_from_line(
-    lines: list[str], index: int
-) -> dict[str, Any] | None:
-    if not _is_thematic_break_line(lines[index]):
-        return None
-    if _is_setext_heading_underline(lines, index):
-        return None
-    block = _AnnotatedSlackBlock({"type": "divider"})
-    block._plain_text = lines[index].strip()
-    return block
-
-
-def _strip_quote_marker(line: str) -> str | None:
-    match = re.match(r"^[ \t]{0,3}>[ \t]?(?P<text>.*)$", line)
-    if not match:
-        return None
-    return match.group("text")
-
-
-def _quote_lines_are_simple(lines: list[str]) -> bool:
-    for line in lines:
-        stripped = line.strip()
-        if not stripped:
-            return False
-        if stripped.startswith(">"):
-            return False
-        if LIST_ITEM_PATTERN.match(stripped):
-            return False
-        if _match_fence_open(stripped):
-            return False
-        if _has_markdown_backslash_escape(stripped):
-            return False
-    return True
-
-
-def _consume_quote_block(
-    lines: list[str], start: int
-) -> tuple[dict[str, Any] | None, int] | None:
-    if _strip_quote_marker(lines[start]) is None:
-        return None
-
-    quote_lines: list[str] = []
-    cursor = start
-    while cursor < len(lines):
-        stripped = _strip_quote_marker(lines[cursor])
-        if stripped is None:
-            break
-        quote_lines.append(stripped)
-        cursor += 1
-
-    if not _quote_lines_are_simple(quote_lines):
-        return None
-
-    quote_text = "\n".join(quote_lines).strip()
-    if not quote_text:
-        return None
-
-    if any(
-        "\n" in quote_text[span_start:span_end]
-        for span_start, span_end in _iter_inline_code_spans(quote_text)
-    ):
-        # A code span crossing quote lines cannot be expressed by the
-        # single-line rich_text tokenizer; the whole quote region falls back
-        # to the markdown path (where Slack renders the span as code) —
-        # returning it line by line would let a later line re-promote alone.
-        return None, cursor
-
-    block = _create_rich_text_block(
-        [
-            {
-                "type": "rich_text_quote",
-                "elements": _create_rich_text_inline_elements(quote_text),
-            }
-        ],
-        plain_text="\n".join(lines[start:cursor]),
-    )
-    return block, cursor
-
-
-def _parse_simple_list_item(line: str) -> dict[str, Any] | None:
-    if _is_thematic_break_line(line):
-        return None
-
-    match = LIST_ITEM_PATTERN.match(line)
-    if not match:
-        return None
-
-    text = line[match.end() :].rstrip()
-    if TASK_LIST_MARKER_PATTERN.match(text):
-        return None
-
-    indent = match.group("indent") or ""
-    if _is_ambiguous_rich_list_indent(indent):
-        return None
-    if _has_markdown_backslash_escape(text):
-        return None
-
-    marker = match.group("marker")
-    return {
-        "style": "ordered" if _is_ordered_list_marker(marker) else "bullet",
-        "number": _ordered_list_marker_number(marker),
-        "indent": _list_indent_level(indent),
-        "text": text.strip() or " ",
-    }
-
-
-def _consume_list_block(
-    lines: list[str], start: int
-) -> tuple[dict[str, Any], int] | None:
-    first_entry = _parse_simple_list_item(lines[start])
-    if first_entry is None:
-        return None
-
-    first_match = LIST_ITEM_PATTERN.match(lines[start])
-    first_indent = _indent_width(first_match.group("indent") if first_match else "")
-    if first_indent > 3:
-        return None
-
-    if start > 0 and lines[start - 1].strip():
-        return None
-
-    entries: list[dict[str, Any]] = []
-    cursor = start
-    while cursor < len(lines) and lines[cursor].strip():
-        entry = _parse_simple_list_item(lines[cursor])
-        if entry is None:
-            return None
-        entries.append(entry)
-        cursor += 1
-
-    if not entries:
-        return None
-
-    lookahead = cursor
-    while lookahead < len(lines) and not lines[lookahead].strip():
-        lookahead += 1
-    if lookahead < len(lines):
-        next_line = lines[lookahead]
-        if _indent_width(next_line) > 0 and _parse_simple_list_item(next_line) is None:
-            return None
-
-    rich_elements: list[dict[str, Any]] = []
-    current_group: dict[str, Any] | None = None
-
-    def flush_group() -> None:
-        nonlocal current_group
-        if current_group:
-            rich_elements.append(current_group)
-        current_group = None
-
-    for entry in entries:
-        group_key = (entry["style"], entry["indent"])
-        if current_group is None or current_group["_key"] != group_key:
-            flush_group()
-            current_group = {
-                "type": "rich_text_list",
-                "style": entry["style"],
-                "indent": entry["indent"],
-                "elements": [],
-                "_key": group_key,
-            }
-            if entry["style"] == "ordered" and entry["number"] > 1:
-                current_group["offset"] = entry["number"] - 1
-
-        current_group["elements"].append(_create_rich_text_section(entry["text"]))
-
-    flush_group()
-    for element in rich_elements:
-        element.pop("_key", None)
-
-    return (
-        _create_rich_text_block(
-            rich_elements,
-            plain_text="\n".join(lines[start:cursor]),
-        ),
-        cursor,
-    )
-
-
-def _find_fence_close_index(
-    lines: list[str], start: int, fence: tuple[str, int]
-) -> int | None:
-    cursor = start + 1
-    while cursor < len(lines):
-        if _is_fence_close(lines[cursor], fence):
-            return cursor
-        cursor += 1
-    return None
-
-
-def _consume_fenced_code_block(
-    lines: list[str], start: int
-) -> tuple[dict[str, Any], int] | None:
-    open_match = FENCE_OPEN_PATTERN.match(lines[start])
-    if not open_match:
-        return None
-
-    fence = _match_fence_open(lines[start])
-    if fence is None:
-        return None
-
-    close_index = _find_fence_close_index(lines, start, fence)
-    if close_index is None:
-        return None
-
-    info = (open_match.group(2) or "").strip()
-    language = info.split()[0] if info else ""
-    code_text = "\n".join(lines[start + 1 : close_index])
-    preformatted: dict[str, Any] = {
-        "type": "rich_text_preformatted",
-        "elements": [{"type": "text", "text": code_text}],
-    }
-    if language and re.match(r"^[A-Za-z0-9_+.#-]+$", language):
-        preformatted["language"] = language
-    return (
-        _create_rich_text_block(
-            [preformatted],
-            plain_text="\n".join(lines[start : close_index + 1]),
-        ),
-        close_index + 1,
-    )
-
-
-def _consume_rich_markdown_block(
-    lines: list[str], index: int
-) -> tuple[dict[str, Any] | None, int] | None:
-    """Consume one promotable construct starting at ``index``.
-
-    Returns ``None`` when nothing was consumed, or ``(block, next_index)``.
-    ``block`` may be ``None`` when the construct was recognized but must stay
-    on the markdown path as a whole region (the caller buffers the raw
-    lines)."""
-    if not lines[index].strip():
-        return None
-
-    consumers = (
-        lambda: _consume_fenced_code_block(lines, index),
-        lambda: (
-            (block, index + 1)
-            if (block := _create_image_block_from_line(lines[index]))
-            else None
-        ),
-        lambda: (
-            (block, index + 1)
-            if (block := _create_divider_block_from_line(lines, index))
-            else None
-        ),
-        lambda: _consume_quote_block(lines, index),
-        lambda: _consume_list_block(lines, index),
-    )
-
-    for consumer in consumers:
-        consumed = consumer()
-        if consumed:
-            return consumed
-    return None
 
 
 def _convert_markdown_text_segment_to_blocks(
@@ -2749,90 +339,6 @@ def convert_markdown_to_slack_blocks(
 convert_markdown_text_to_blocks = convert_markdown_to_slack_blocks
 
 
-def _block_expansion_weight(block: dict[str, Any]) -> int:
-    """Weight of one block against Slack's per-message expansion budget.
-
-    Slack expands ``markdown`` blocks server-side and enforces the 50-item
-    limit on the expanded result, so a markdown block counts as its estimated
-    expansion; every other block type posts as a single item.
-    """
-    if not isinstance(block, dict) or block.get("type") != "markdown":
-        return 1
-    annotated = getattr(block, "_expansion_items", None)
-    if isinstance(annotated, int) and annotated > 0:
-        return annotated
-    return _estimate_markdown_expansion_items(str(block.get("text", "")))
-
-
-_TEXT_SIZE_KEYS = frozenset({"text", "url", "alt_text", "image_url"})
-
-
-def _block_text_size(value: Any) -> int:
-    """Rough text payload of a block against the per-message total budget.
-
-    Slack's ``msg_blocks_too_long`` check counts content across block types
-    (a 11,900-char markdown block plus a 1,400-char rich_text was rejected),
-    so this sums every string under content-carrying keys, recursively.
-    """
-    if isinstance(value, dict):
-        total = 0
-        for key, sub in value.items():
-            if key in _TEXT_SIZE_KEYS and isinstance(sub, str):
-                total += len(sub)
-            else:
-                total += _block_text_size(sub)
-        return total
-    if isinstance(value, list):
-        return sum(_block_text_size(item) for item in value)
-    return 0
-
-
-def split_blocks_by_table(blocks: list[dict[str, Any]]) -> list[list[dict[str, Any]]]:
-    """Split blocks to satisfy Slack table and per-message constraints.
-
-    A message holds at most one ``table`` block, at most
-    ``SLACK_MAX_BLOCKS_PER_MESSAGE`` posted blocks, at most
-    ``SLACK_MAX_EXPANSION_ITEMS_PER_MESSAGE`` estimated post-expansion items
-    (headings and thematic breaks inside ``markdown`` blocks count
-    individually toward that budget), and at most
-    ``SLACK_MAX_MESSAGE_BLOCKS_TEXT_LENGTH`` characters of block text in
-    total across block types.
-    """
-    messages: list[list[dict[str, Any]]] = []
-    current_message: list[dict[str, Any]] = []
-    current_weight = 0
-    current_text_size = 0
-
-    for block in blocks or []:
-        if isinstance(block, dict) and block.get("type") == "table":
-            if current_message:
-                messages.append(current_message)
-            messages.append([block])
-            current_message = []
-            current_weight = 0
-            current_text_size = 0
-        else:
-            weight = _block_expansion_weight(block)
-            text_size = _block_text_size(block)
-            if current_message and (
-                len(current_message) >= SLACK_MAX_BLOCKS_PER_MESSAGE
-                or current_weight + weight > SLACK_MAX_EXPANSION_ITEMS_PER_MESSAGE
-                or current_text_size + text_size > _MESSAGE_BLOCKS_TEXT_TARGET
-            ):
-                messages.append(current_message)
-                current_message = []
-                current_weight = 0
-                current_text_size = 0
-            current_message.append(block)
-            current_weight += weight
-            current_text_size += text_size
-
-    if current_message:
-        messages.append(current_message)
-
-    return messages
-
-
 def convert_markdown_to_slack_messages(
     markdown_text: str,
     *,
@@ -2860,121 +366,3 @@ def convert_markdown_to_slack_payloads(
         fallback_text = build_fallback_text_from_blocks(blocks).strip()
         payloads.append({"blocks": blocks, "text": fallback_text or " "})
     return payloads
-
-
-def _markdown_block_to_plain_text(block: dict[str, Any]) -> str:
-    """Downgrade a ``markdown`` block, preferring the build-time annotation."""
-    text = getattr(block, "_plain_text", None) or ""
-    if text:
-        return str(text)
-    raw_text = block.get("text", "")
-    if not raw_text:
-        return ""
-    return _normalize_markdown_block_plain_text(
-        _strip_synthetic_blank_line_placeholders(
-            _strip_synthetic_spaces_from_plain_text(
-                strip_zero_width_spaces(raw_text),
-                getattr(block, "_synthetic_space_indices", None),
-            ),
-            getattr(block, "_synthetic_blank_line_indices", None),
-        )
-    )
-
-
-def _blocks_to_downgrade_parts(
-    blocks: list[dict[str, Any]], *, fallback: bool
-) -> list[str]:
-    """Shared block walker behind :func:`blocks_to_plain_text` and
-    :func:`build_fallback_text_from_blocks`.
-
-    The two public functions intentionally differ in a few policies, kept
-    explicit on the ``fallback`` flag: fallback keeps table cells verbatim
-    (empty cells preserve column alignment) and emits a whole table as one
-    part, while the plain-text view strips zero-width spaces, drops empty
-    cells, emits one part per row, and surfaces ``text`` from unknown blocks.
-    """
-    parts: list[str] = []
-
-    for block in blocks or []:
-        if not isinstance(block, dict):
-            continue
-        block_type = block.get("type")
-
-        if block_type == "markdown":
-            text = _markdown_block_to_plain_text(block)
-            if text.strip() if fallback else text:
-                parts.append(text)
-        elif block_type == "table":
-            row_texts: list[str] = []
-            for row in block.get("rows") or []:
-                if not isinstance(row, list):
-                    continue
-                if fallback:
-                    cell_texts = [
-                        extract_plain_text_from_table_cell(cell) for cell in row
-                    ]
-                else:
-                    cell_texts = []
-                    for cell in row:
-                        cell_text = extract_plain_text_from_table_cell(cell)
-                        if cell_text:
-                            cell_texts.append(strip_zero_width_spaces(cell_text))
-                if cell_texts:
-                    row_texts.append(" | ".join(cell_texts))
-            if fallback:
-                if row_texts:
-                    parts.append("\n".join(row_texts))
-            else:
-                parts.extend(row_texts)
-        elif block_type == "rich_text":
-            text = _rich_text_block_to_plain_text(block)
-            if text.strip() if fallback else text:
-                parts.append(text)
-        elif block_type == "header":
-            text = block.get("text", {})
-            if isinstance(text, dict) and text.get("text"):
-                parts.append(str(text.get("text", "")))
-        elif block_type == "image":
-            alt_text = str(block.get("alt_text", "")).strip()
-            image_url = str(block.get("image_url", "")).strip()
-            image_text = alt_text or image_url
-            if alt_text and image_url:
-                image_text = f"{alt_text} ({image_url})"
-            if image_text:
-                parts.append(image_text)
-        elif block_type == "divider":
-            parts.append(getattr(block, "_plain_text", None) or "---")
-        elif not fallback:
-            text = block.get("text", "")
-            if isinstance(text, dict):
-                # e.g. a ``section`` block carries ``text: {type, text}``.
-                text = text.get("text", "")
-            if text:
-                parts.append(str(text))
-
-    return parts
-
-
-def blocks_to_plain_text(blocks: list[dict[str, Any]]) -> str:
-    """Build plain text representation from Slack blocks."""
-    parts = _blocks_to_downgrade_parts(blocks, fallback=False)
-    return "\n".join([p for p in parts if p]).strip()
-
-
-def build_fallback_text_from_blocks(blocks: list[dict[str, Any]]) -> str:
-    """Build Slack fallback text from block structure."""
-    parts = _blocks_to_downgrade_parts(blocks, fallback=True)
-    return "\n\n".join([part for part in parts if part.strip()])
-
-
-# Backward-compatible helper retained for existing imports.
-def parse_markdown_table(table_text: str) -> list[list[str]]:
-    """Parse markdown table into row/cell text matrix."""
-    rows: list[list[str]] = []
-    for line in [line for line in table_text.strip().splitlines() if line.strip()]:
-        if TABLE_SEPARATOR_PATTERN.match(line.strip()):
-            continue
-        if "|" not in line:
-            continue
-        rows.append(_split_markdown_table_cells(line))
-    return rows
